Backed out 8 changesets (bug 1672520) for Build bustage in gecko/netwerk/protocol/http/nsHttpConnectionMgr.cpp. CLOSED TREE
Backed out changeset fd7e69c065d9 (bug 1672520)
Backed out changeset dd9c9c3c959c (bug 1672520)
Backed out changeset 84ad3077d7ef (bug 1672520)
Backed out changeset 09de9356d225 (bug 1672520)
Backed out changeset 6a0a89a02779 (bug 1672520)
Backed out changeset c1b12059f317 (bug 1672520)
Backed out changeset 473498b3df4c (bug 1672520)
Backed out changeset e2a2ecf84209 (bug 1672520)
