+++ netwerk/protocol/http/moz.build
@@ -93,10 +93,7 @@ UNIFIED_SOURCES += [
     "ClassifierDummyChannelChild.cpp",
     "ClassifierDummyChannelParent.cpp",
     "ConnectionDiagnostics.cpp",
-    "ConnectionEntry.cpp",
-    "ConnectionHandle.cpp",
     "DelayHttpChannelQueue.cpp",
-    "HalfOpenSocket.cpp",
     "Http2Compression.cpp",
     "Http2Push.cpp",
     "Http2Session.cpp",
@@ -141,8 +138,6 @@ UNIFIED_SOURCES += [
     "NullHttpChannel.cpp",
     "NullHttpTransaction.cpp",
     "ParentChannelListener.cpp",
-    "PendingTransactionInfo.cpp",
-    "PendingTransactionQueue.cpp",
     "QuicSocketControl.cpp",
     "SpeculativeTransaction.cpp",
     "TRRServiceChannel.cpp",
