+++ netwerk/protocol/http/nsHttpConnectionMgr.cpp
@@ -16,6 +16,7 @@
 #include <utility>
 
 #include "NullHttpTransaction.h"
+#include "mozilla/ChaosMode.h"
 #include "mozilla/Services.h"
 #include "mozilla/Telemetry.h"
 #include "mozilla/Unused.h"
@@ -39,7 +40,6 @@
 #include "nsNetCID.h"
 #include "nsNetUtil.h"
 #include "nsQueryObject.h"
-#include "ConnectionHandle.h"
 #include "HttpConnectionUDP.h"
 #include "SpeculativeTransaction.h"
 #include "TCPFastOpenLayer.h"
@@ -50,6 +50,68 @@ namespace net {
 //-----------------------------------------------------------------------------
 
 NS_IMPL_ISUPPORTS(nsHttpConnectionMgr, nsIObserver)
+
+// This function decides the transaction's order in the pending queue.
+// Given two transactions t1 and t2, returning true means that t2 is
+// more important than t1 and thus should be dispatched first.
+static bool TransactionComparator(nsHttpTransaction* t1,
+                                  nsHttpTransaction* t2) {
+  bool t1Blocking =
+      t1->Caps() & (NS_HTTP_LOAD_AS_BLOCKING | NS_HTTP_LOAD_UNBLOCKED);
+  bool t2Blocking =
+      t2->Caps() & (NS_HTTP_LOAD_AS_BLOCKING | NS_HTTP_LOAD_UNBLOCKED);
+
+  if (t1Blocking > t2Blocking) {
+    return false;
+  }
+
+  if (t2Blocking > t1Blocking) {
+    return true;
+  }
+
+  return t1->Priority() >= t2->Priority();
+}
+
+void nsHttpConnectionMgr::InsertTransactionSorted(
+    nsTArray<RefPtr<nsHttpConnectionMgr::PendingTransactionInfo>>& pendingQ,
+    nsHttpConnectionMgr::PendingTransactionInfo* pendingTransInfo,
+    bool aInsertAsFirstForTheSamePriority /*= false*/) {
+  // insert the transaction into the front of the queue based on following
+  // rules:
+  // 1. The transaction has NS_HTTP_LOAD_AS_BLOCKING or NS_HTTP_LOAD_UNBLOCKED.
+  // 2. The transaction's priority is higher.
+  //
+  // search in reverse order under the assumption that many of the
+  // existing transactions will have the same priority (usually 0).
+
+  nsHttpTransaction* trans = pendingTransInfo->mTransaction;
+
+  for (int32_t i = pendingQ.Length() - 1; i >= 0; --i) {
+    nsHttpTransaction* t = pendingQ[i]->mTransaction;
+    if (TransactionComparator(trans, t)) {
+      if (ChaosMode::isActive(ChaosFeature::NetworkScheduling) ||
+          aInsertAsFirstForTheSamePriority) {
+        int32_t samePriorityCount;
+        for (samePriorityCount = 0; i - samePriorityCount >= 0;
+             ++samePriorityCount) {
+          if (pendingQ[i - samePriorityCount]->mTransaction->Priority() !=
+              trans->Priority()) {
+            break;
+          }
+        }
+        if (aInsertAsFirstForTheSamePriority) {
+          i -= samePriorityCount;
+        } else {
+          // skip over 0...all of the elements with the same priority.
+          i -= ChaosMode::randomUint32LessThan(samePriorityCount + 1);
+        }
+      }
+      pendingQ.InsertElementAt(i + 1, pendingTransInfo);
+      return;
+    }
+  }
+  pendingQ.InsertElementAt(0, pendingTransInfo);
+}
 
 //-----------------------------------------------------------------------------
 
@@ -591,11 +653,11 @@ void nsHttpConnectionMgr::OnMsgClearConnectionHistory(int32_t,
   LOG(("nsHttpConnectionMgr::OnMsgClearConnectionHistory"));
 
   for (auto iter = mCT.Iter(); !iter.Done(); iter.Next()) {
-    RefPtr<ConnectionEntry> ent = iter.Data();
-    if (ent->IdleConnectionsLength() == 0 && ent->ActiveConnsLength() == 0 &&
-        ent->HalfOpensLength() == 0 && ent->UrgentStartQueueLength() == 0 &&
-        ent->PendingQueueLength() == 0 &&
-        ent->HalfOpenFastOpenBackupsLength() == 0 && !ent->mDoNotDestroy) {
+    RefPtr<nsConnectionEntry> ent = iter.Data();
+    if (ent->mIdleConns.Length() == 0 && ent->mActiveConns.Length() == 0 &&
+        ent->mHalfOpens.Length() == 0 && ent->mUrgentStartQ.Length() == 0 &&
+        ent->PendingQLength() == 0 &&
+        ent->mHalfOpenFastOpenBackups.Length() == 0 && !ent->mDoNotDestroy) {
       iter.Remove();
     }
   }
@@ -609,12 +671,14 @@ nsresult nsHttpConnectionMgr::CloseIdleConnection(nsHttpConnection* conn) {
     return NS_ERROR_UNEXPECTED;
   }
 
-  ConnectionEntry* ent = mCT.GetWeak(conn->ConnectionInfo()->HashKey());
+  nsConnectionEntry* ent = mCT.GetWeak(conn->ConnectionInfo()->HashKey());
 
-  if (!ent || NS_FAILED(ent->CloseIdleConnection(conn))) {
-    return NS_ERROR_UNEXPECTED;
-  }
+  RefPtr<nsHttpConnection> deleteProtector(conn);
+  if (!ent || !ent->mIdleConns.RemoveElement(conn)) return NS_ERROR_UNEXPECTED;
 
+  conn->Close(NS_ERROR_ABORT);
+  mNumIdleConns--;
+  ConditionallyStopPruneDeadConnectionsTimer();
   return NS_OK;
 }
 
@@ -627,17 +691,19 @@ nsresult nsHttpConnectionMgr::RemoveIdleConnection(nsHttpConnection* conn) {
     return NS_ERROR_UNEXPECTED;
   }
 
-  ConnectionEntry* ent = mCT.GetWeak(conn->ConnectionInfo()->HashKey());
+  nsConnectionEntry* ent = mCT.GetWeak(conn->ConnectionInfo()->HashKey());
 
-  if (!ent || NS_FAILED(ent->RemoveIdleConnection(conn))) {
+  if (!ent || !ent->mIdleConns.RemoveElement(conn)) {
     return NS_ERROR_UNEXPECTED;
   }
 
+  mNumIdleConns--;
+  ConditionallyStopPruneDeadConnectionsTimer();
   return NS_OK;
 }
 
 HttpConnectionBase* nsHttpConnectionMgr::FindCoalescableConnectionByHashKey(
-    ConnectionEntry* ent, const nsCString& key, bool justKidding,
+    nsConnectionEntry* ent, const nsCString& key, bool justKidding,
     bool aNoHttp3) {
   MOZ_ASSERT(OnSocketThread(), "not on socket thread");
   MOZ_ASSERT(ent->mConnInfo);
@@ -722,7 +788,7 @@ static void BuildOriginFrameHashKey(nsACString& newKey,
 }
 
 HttpConnectionBase* nsHttpConnectionMgr::FindCoalescableConnection(
-    ConnectionEntry* ent, bool justKidding, bool aNoHttp3) {
+    nsConnectionEntry* ent, bool justKidding, bool aNoHttp3) {
   MOZ_ASSERT(OnSocketThread(), "not on socket thread");
   MOZ_ASSERT(ent->mConnInfo);
   nsHttpConnectionInfo* ci = ent->mConnInfo;
@@ -757,7 +823,7 @@ HttpConnectionBase* nsHttpConnectionMgr::FindCoalescableConnection(
 }
 
 void nsHttpConnectionMgr::UpdateCoalescingForNewConn(
-    HttpConnectionBase* newConn, ConnectionEntry* ent) {
+    HttpConnectionBase* newConn, nsConnectionEntry* ent) {
   MOZ_ASSERT(OnSocketThread(), "not on socket thread");
   MOZ_ASSERT(newConn);
   MOZ_ASSERT(newConn->ConnectionInfo());
@@ -808,13 +874,45 @@ void nsHttpConnectionMgr::UpdateCoalescingForNewConn(
         do_GetWeakReference(static_cast<nsISupportsWeakReference*>(newConn)));
   }
 
-  // this is a new connection that can be coalesced onto. hooray!
-  // if there are other connection to this entry (e.g.
-  // some could still be handshaking, shutting down, etc..) then close
-  // them down after any transactions that are on them are complete.
-  // This probably happened due to the parallel connection algorithm
-  // that is used only before the host is known to speak h2.
-  ent->MakeAllDontReuseExcept(newConn);
+  // Cancel any other pending connections - their associated transactions
+  // are in the pending queue and will be dispatched onto this new connection
+  for (int32_t index = ent->mHalfOpens.Length() - 1; index >= 0; --index) {
+    RefPtr<nsHalfOpenSocket> half = ent->mHalfOpens[index];
+    LOG(("UpdateCoalescingForNewConn() forcing halfopen abandon %p\n",
+         half.get()));
+    ent->mHalfOpens[index]->Abandon();
+  }
+
+  if (ent->mActiveConns.Length() > 1) {
+    // this is a new connection that can be coalesced onto. hooray!
+    // if there are other connection to this entry (e.g.
+    // some could still be handshaking, shutting down, etc..) then close
+    // them down after any transactions that are on them are complete.
+    // This probably happened due to the parallel connection algorithm
+    // that is used only before the host is known to speak h2.
+    for (uint32_t index = 0; index < ent->mActiveConns.Length(); ++index) {
+      HttpConnectionBase* otherConn = ent->mActiveConns[index];
+      if (otherConn != newConn) {
+        LOG(
+            ("UpdateCoalescingForNewConn() shutting down old connection (%p) "
+             "because new "
+             "spdy connection (%p) takes precedence\n",
+             otherConn, newConn));
+        otherConn->DontReuse();
+      }
+    }
+  }
+
+  for (int32_t index = ent->mHalfOpenFastOpenBackups.Length() - 1; index >= 0;
+       --index) {
+    LOG(
+        ("UpdateCoalescingForNewConn() shutting down connection in fast "
+         "open state (%p) because new spdy connection (%p) takes "
+         "precedence\n",
+         ent->mHalfOpenFastOpenBackups[index].get(), newConn));
+    RefPtr<nsHalfOpenSocket> half = ent->mHalfOpenFastOpenBackups[index];
+    half->CancelFastOpenConnection();
+  }
 }
 
 // This function lets a connection, after completing the NPN phase,
@@ -830,7 +928,7 @@ void nsHttpConnectionMgr::ReportSpdyConnection(nsHttpConnection* conn,
   if (!conn->ConnectionInfo()) {
     return;
   }
-  ConnectionEntry* ent = mCT.GetWeak(conn->ConnectionInfo()->HashKey());
+  nsConnectionEntry* ent = mCT.GetWeak(conn->ConnectionInfo()->HashKey());
   if (!ent || !usingSpdy) {
     return;
   }
@@ -868,7 +966,7 @@ void nsHttpConnectionMgr::ReportHttp3Connection(HttpConnectionBase* conn) {
   if (!conn->ConnectionInfo()) {
     return;
   }
-  ConnectionEntry* ent = mCT.GetWeak(conn->ConnectionInfo()->HashKey());
+  nsConnectionEntry* ent = mCT.GetWeak(conn->ConnectionInfo()->HashKey());
   if (!ent) {
     return;
   }
@@ -894,8 +992,8 @@ void nsHttpConnectionMgr::ReportHttp3Connection(HttpConnectionBase* conn) {
 
 //-----------------------------------------------------------------------------
 bool nsHttpConnectionMgr::DispatchPendingQ(
-    nsTArray<RefPtr<PendingTransactionInfo>>& pendingQ, ConnectionEntry* ent,
-    bool considerAll) {
+    nsTArray<RefPtr<nsHttpConnectionMgr::PendingTransactionInfo>>& pendingQ,
+    nsConnectionEntry* ent, bool considerAll) {
   MOZ_ASSERT(OnSocketThread(), "not on socket thread");
 
   PendingTransactionInfo* pendingTransInfo = nullptr;
@@ -907,14 +1005,62 @@ bool nsHttpConnectionMgr::DispatchPendingQ(
   // dispatch. if considerAll == true then try and dispatch all items.
   for (uint32_t i = 0; i < pendingQ.Length();) {
     pendingTransInfo = pendingQ[i];
+    LOG((
+        "nsHttpConnectionMgr::DispatchPendingQ "
+        "[trans=%p, halfOpen=%p, activeConn=%p]\n",
+        pendingTransInfo->mTransaction.get(), pendingTransInfo->mHalfOpen.get(),
+        pendingTransInfo->mActiveConn.get()));
 
-    bool alreadyHalfOpenOrWaitingForTLS =
-        pendingTransInfo->IsAlreadyClaimedInitializingConn();
+    // When this transaction has already established a half-open
+    // connection, we want to prevent any duplicate half-open
+    // connections from being established and bound to this
+    // transaction. Allow only use of an idle persistent connection
+    // (if found) for transactions referred by a half-open connection.
+    bool alreadyHalfOpenOrWaitingForTLS = false;
+    if (pendingTransInfo->mHalfOpen) {
+      MOZ_ASSERT(!pendingTransInfo->mActiveConn);
+      RefPtr<nsHalfOpenSocket> halfOpen =
+          do_QueryReferent(pendingTransInfo->mHalfOpen);
+      LOG(
+          ("nsHttpConnectionMgr::DispatchPendingQ "
+           "[trans=%p, halfOpen=%p]\n",
+           pendingTransInfo->mTransaction.get(), halfOpen.get()));
+      if (halfOpen) {
+        alreadyHalfOpenOrWaitingForTLS = true;
+      } else {
+        // If we have not found the halfOpen socket, remove the pointer.
+        pendingTransInfo->mHalfOpen = nullptr;
+      }
+    } else if (pendingTransInfo->mActiveConn) {
+      MOZ_ASSERT(!pendingTransInfo->mHalfOpen);
+      RefPtr<HttpConnectionBase> activeConn =
+          do_QueryReferent(pendingTransInfo->mActiveConn);
+      LOG(
+          ("nsHttpConnectionMgr::DispatchPendingQ "
+           "[trans=%p, activeConn=%p]\n",
+           pendingTransInfo->mTransaction.get(), activeConn.get()));
+      // Check if this transaction claimed a connection that is still
+      // performing tls handshake with a NullHttpTransaction or it is between
+      // finishing tls and reclaiming (When nullTrans finishes tls handshake,
+      // httpConnection does not have a transaction any more and a
+      // ReclaimConnection is dispatched). But if an error occurred the
+      // connection will be closed, it will exist but CanReused will be
+      // false.
+      if (activeConn &&
+          ((activeConn->Transaction() &&
+            activeConn->Transaction()->IsNullTransaction()) ||
+           (!activeConn->Transaction() && activeConn->CanReuse()))) {
+        alreadyHalfOpenOrWaitingForTLS = true;
+      } else {
+        // If we have not found the connection, remove the pointer.
+        pendingTransInfo->mActiveConn = nullptr;
+      }
+    }
 
     rv = TryDispatchTransaction(
         ent,
         alreadyHalfOpenOrWaitingForTLS ||
-            !!pendingTransInfo->Transaction()->TunnelProvider(),
+            !!pendingTransInfo->mTransaction->TunnelProvider(),
         pendingTransInfo);
     if (NS_SUCCEEDED(rv) || (rv != NS_ERROR_NOT_AVAILABLE)) {
       if (NS_SUCCEEDED(rv)) {
@@ -925,6 +1071,7 @@ bool nsHttpConnectionMgr::DispatchPendingQ(
              "TryDispatchTransaction returning hard error %" PRIx32 "\n",
              static_cast<uint32_t>(rv)));
       }
+      ReleaseClaimedSockets(ent, pendingTransInfo);
       if (pendingQ.RemoveElement(pendingTransInfo)) {
         // pendingTransInfo is now potentially destroyed
         dispatchedSuccessfully = true;
@@ -941,8 +1088,18 @@ bool nsHttpConnectionMgr::DispatchPendingQ(
   return dispatchedSuccessfully;
 }
 
+uint32_t nsHttpConnectionMgr::TotalActiveConnections(
+    nsConnectionEntry* ent) const {
+  // Add in the in-progress tcp connections, we will assume they are
+  // keepalive enabled.
+  // Exclude half-open's that has already created a usable connection.
+  // This prevents the limit being stuck on ipv6 connections that
+  // eventually time out after typical 21 seconds of no ACK+SYN reply.
+  return ent->mActiveConns.Length() + ent->UnconnectedHalfOpens();
+}
+
 uint32_t nsHttpConnectionMgr::MaxPersistConnections(
-    ConnectionEntry* ent) const {
+    nsConnectionEntry* ent) const {
   if (ent->mConnInfo->UsingHttpProxy() && !ent->mConnInfo->UsingConnect()) {
     return static_cast<uint32_t>(mMaxPersistConnsPerProxy);
   }
@@ -951,13 +1108,13 @@ uint32_t nsHttpConnectionMgr::MaxPersistConnections(
 }
 
 void nsHttpConnectionMgr::PreparePendingQForDispatching(
-    ConnectionEntry* ent, nsTArray<RefPtr<PendingTransactionInfo>>& pendingQ,
+    nsConnectionEntry* ent, nsTArray<RefPtr<PendingTransactionInfo>>& pendingQ,
     bool considerAll) {
   MOZ_ASSERT(OnSocketThread(), "not on socket thread");
 
   pendingQ.Clear();
 
-  uint32_t totalCount = ent->TotalActiveConnections();
+  uint32_t totalCount = TotalActiveConnections(ent);
   uint32_t maxPersistConns = MaxPersistConnections(ent);
   uint32_t availableConnections =
       maxPersistConns > totalCount ? maxPersistConns - totalCount : 0;
@@ -1034,7 +1191,7 @@ void nsHttpConnectionMgr::PreparePendingQForDispatching(
   pendingQ.AppendElements(std::move(remainingPendingQ));
 }
 
-bool nsHttpConnectionMgr::ProcessPendingQForEntry(ConnectionEntry* ent,
+bool nsHttpConnectionMgr::ProcessPendingQForEntry(nsConnectionEntry* ent,
                                                   bool considerAll) {
   MOZ_ASSERT(OnSocketThread(), "not on socket thread");
 
@@ -1042,29 +1199,71 @@ bool nsHttpConnectionMgr::ProcessPendingQForEntry(ConnectionEntry* ent,
       ("nsHttpConnectionMgr::ProcessPendingQForEntry "
        "[ci=%s ent=%p active=%zu idle=%zu urgent-start-queue=%zu"
        " queued=%zu]\n",
-       ent->mConnInfo->HashKey().get(), ent, ent->ActiveConnsLength(),
-       ent->IdleConnectionsLength(), ent->UrgentStartQueueLength(),
-       ent->PendingQueueLength()));
+       ent->mConnInfo->HashKey().get(), ent, ent->mActiveConns.Length(),
+       ent->mIdleConns.Length(), ent->mUrgentStartQ.Length(),
+       ent->PendingQLength()));
 
   if (LOG_ENABLED()) {
-    ent->PrintPendingQ();
-    ent->LogConnections();
+    LOG(("urgent queue ["));
+    for (const auto& info : ent->mUrgentStartQ) {
+      LOG(("  %p", info->mTransaction.get()));
+    }
+    for (auto it = ent->mPendingTransactionTable.Iter(); !it.Done();
+         it.Next()) {
+      LOG(("] window id = %" PRIx64 " queue [", it.Key()));
+      for (const auto& info : *it.UserData()) {
+        LOG(("  %p", info->mTransaction.get()));
+      }
+    }
+    if (!ent->mConnInfo->IsHttp3()) {
+      LOG(("] active urgent conns ["));
+      for (HttpConnectionBase* conn : ent->mActiveConns) {
+        RefPtr<nsHttpConnection> connTCP = do_QueryObject(conn);
+        MOZ_ASSERT(connTCP);
+        if (connTCP->IsUrgentStartPreferred()) {
+          LOG(("  %p", conn));
+        }
+      }
+      LOG(("] active regular conns ["));
+      for (HttpConnectionBase* conn : ent->mActiveConns) {
+        RefPtr<nsHttpConnection> connTCP = do_QueryObject(conn);
+        MOZ_ASSERT(connTCP);
+        if (!connTCP->IsUrgentStartPreferred()) {
+          LOG(("  %p", conn));
+        }
+      }
+
+      LOG(("] idle urgent conns ["));
+      for (nsHttpConnection* conn : ent->mIdleConns) {
+        if (conn->IsUrgentStartPreferred()) {
+          LOG(("  %p", conn));
+        }
+      }
+      LOG(("] idle regular conns ["));
+      for (nsHttpConnection* conn : ent->mIdleConns) {
+        if (!conn->IsUrgentStartPreferred()) {
+          LOG(("  %p", conn));
+        }
+      }
+    } else {
+      for (HttpConnectionBase* conn : ent->mActiveConns) {
+        LOG(("  %p", conn));
+      }
+      MOZ_ASSERT(ent->mIdleConns.Length() == 0);
+    }
+    LOG(("]"));
   }
 
-  if (!ent->PendingQueueLength() && !ent->UrgentStartQueueLength()) {
+  if (!ent->mUrgentStartQ.Length() && !ent->PendingQLength()) {
     return false;
   }
   ProcessSpdyPendingQ(ent);
 
   bool dispatchedSuccessfully = false;
 
-  if (ent->UrgentStartQueueLength()) {
-    nsTArray<RefPtr<PendingTransactionInfo>> pendingQ;
-    ent->AppendPendingUrgentStartQ(pendingQ);
-    dispatchedSuccessfully = DispatchPendingQ(pendingQ, ent, considerAll);
-    for (const auto& transactionInfo : Reversed(pendingQ)) {
-      ent->InsertTransaction(transactionInfo);
-    }
+  if (!ent->mUrgentStartQ.IsEmpty()) {
+    dispatchedSuccessfully =
+        DispatchPendingQ(ent->mUrgentStartQ, ent, considerAll);
   }
 
   if (dispatchedSuccessfully && !considerAll) {
@@ -1099,7 +1298,7 @@ bool nsHttpConnectionMgr::ProcessPendingQForEntry(ConnectionEntry* ent,
 bool nsHttpConnectionMgr::ProcessPendingQForEntry(nsHttpConnectionInfo* ci) {
   MOZ_ASSERT(OnSocketThread(), "not on socket thread");
 
-  ConnectionEntry* ent = mCT.GetWeak(ci->HashKey());
+  nsConnectionEntry* ent = mCT.GetWeak(ci->HashKey());
   if (ent) return ProcessPendingQForEntry(ent, false);
   return false;
 }
@@ -1109,10 +1308,10 @@ bool nsHttpConnectionMgr::ProcessPendingQForEntry(nsHttpConnectionInfo* ci) {
 //  (1) at max-connections
 //  (2) keep-alive enabled and at max-persistent-connections-per-server/proxy
 //  (3) keep-alive disabled and at max-connections-per-server
-bool nsHttpConnectionMgr::AtActiveConnectionLimit(ConnectionEntry* ent,
+bool nsHttpConnectionMgr::AtActiveConnectionLimit(nsConnectionEntry* ent,
                                                   uint32_t caps) {
   nsHttpConnectionInfo* ci = ent->mConnInfo;
-  uint32_t totalCount = ent->TotalActiveConnections();
+  uint32_t totalCount = TotalActiveConnections(ent);
 
   if (ci->IsHttp3()) {
     return totalCount > 0;
@@ -1158,27 +1357,139 @@ bool nsHttpConnectionMgr::AtActiveConnectionLimit(ConnectionEntry* ent,
   return result;
 }
 
+void nsHttpConnectionMgr::ClosePersistentConnections(nsConnectionEntry* ent) {
+  LOG(("nsHttpConnectionMgr::ClosePersistentConnections [ci=%s]\n",
+       ent->mConnInfo->HashKey().get()));
+  while (ent->mIdleConns.Length()) {
+    RefPtr<nsHttpConnection> conn(ent->mIdleConns[0]);
+    ent->mIdleConns.RemoveElementAt(0);
+    mNumIdleConns--;
+    conn->Close(NS_ERROR_ABORT);
+  }
+
+  int32_t activeCount = ent->mActiveConns.Length();
+  for (int32_t i = 0; i < activeCount; i++) ent->mActiveConns[i]->DontReuse();
+  for (int32_t index = ent->mHalfOpenFastOpenBackups.Length() - 1; index >= 0;
+       --index) {
+    RefPtr<nsHalfOpenSocket> half = ent->mHalfOpenFastOpenBackups[index];
+    half->CancelFastOpenConnection();
+  }
+}
+
+bool nsHttpConnectionMgr::RestrictConnections(nsConnectionEntry* ent) {
+  MOZ_ASSERT(OnSocketThread(), "not on socket thread");
+
+  if (ent->AvailableForDispatchNow()) {
+    // this might be a h2/spdy connection in this connection entry that
+    // is able to be immediately muxxed, or it might be one that
+    // was found in the same state through a coalescing hash
+    LOG(
+        ("nsHttpConnectionMgr::RestrictConnections %p %s restricted due to "
+         "active >=h2\n",
+         ent, ent->mConnInfo->HashKey().get()));
+    return true;
+  }
+
+  // If this host is trying to negotiate a SPDY session right now,
+  // don't create any new ssl connections until the result of the
+  // negotiation is known.
+
+  bool doRestrict = ent->mConnInfo->FirstHopSSL() &&
+                    gHttpHandler->IsSpdyEnabled() && ent->mUsingSpdy &&
+                    (ent->mHalfOpens.Length() || ent->mActiveConns.Length());
+
+  // If there are no restrictions, we are done
+  if (!doRestrict) return false;
+
+  // If the restriction is based on a tcp handshake in progress
+  // let that connect and then see if it was SPDY or not
+  if (ent->UnconnectedHalfOpens()) {
+    return true;
+  }
+
+  // There is a concern that a host is using a mix of HTTP/1 and SPDY.
+  // In that case we don't want to restrict connections just because
+  // there is a single active HTTP/1 session in use.
+  if (ent->mUsingSpdy && ent->mActiveConns.Length()) {
+    bool confirmedRestrict = false;
+    for (uint32_t index = 0; index < ent->mActiveConns.Length(); ++index) {
+      HttpConnectionBase* conn = ent->mActiveConns[index];
+      RefPtr<nsHttpConnection> connTCP = do_QueryObject(conn);
+      if ((connTCP && !connTCP->ReportedNPN()) || conn->CanDirectlyActivate()) {
+        confirmedRestrict = true;
+        break;
+      }
+    }
+    doRestrict = confirmedRestrict;
+    if (!confirmedRestrict) {
+      LOG(
+          ("nsHttpConnectionMgr spdy connection restriction to "
+           "%s bypassed.\n",
+           ent->mConnInfo->Origin()));
+    }
+  }
+  return doRestrict;
+}
+
 // returns NS_OK if a connection was started
 // return NS_ERROR_NOT_AVAILABLE if a new connection cannot be made due to
 //        ephemeral limits
 // returns other NS_ERROR on hard failure conditions
 nsresult nsHttpConnectionMgr::MakeNewConnection(
-    ConnectionEntry* ent, PendingTransactionInfo* pendingTransInfo) {
+    nsConnectionEntry* ent, PendingTransactionInfo* pendingTransInfo) {
+  nsHttpTransaction* trans = pendingTransInfo->mTransaction;
+
   LOG(("nsHttpConnectionMgr::MakeNewConnection %p ent=%p trans=%p", this, ent,
-       pendingTransInfo->Transaction()));
+       trans));
   MOZ_ASSERT(OnSocketThread(), "not on socket thread");
 
-  if (ent->FindConnToClaim(pendingTransInfo)) {
-    return NS_OK;
+  uint32_t halfOpenLength = ent->mHalfOpens.Length();
+  for (uint32_t i = 0; i < halfOpenLength; i++) {
+    auto halfOpen = ent->mHalfOpens[i];
+    if (halfOpen->AcceptsTransaction(trans) && halfOpen->Claim()) {
+      // We've found a speculative connection or a connection that
+      // is free to be used in the half open list.
+      // A free to be used connection is a connection that was
+      // open for a concrete transaction, but that trunsaction
+      // ended up using another connection.
+      LOG(
+          ("nsHttpConnectionMgr::MakeNewConnection [ci = %s]\n"
+           "Found a speculative or a free-to-use half open connection\n",
+           ent->mConnInfo->HashKey().get()));
+      pendingTransInfo->mHalfOpen = do_GetWeakReference(
+          static_cast<nsISupportsWeakReference*>(ent->mHalfOpens[i]));
+      // return OK because we have essentially opened a new connection
+      // by converting a speculative half-open to general use
+      return NS_OK;
+    }
   }
 
-  nsHttpTransaction* trans = pendingTransInfo->Transaction();
+  // consider null transactions that are being used to drive the ssl handshake
+  // if the transaction creating this connection can re-use persistent
+  // connections
+  if (trans->Caps() & NS_HTTP_ALLOW_KEEPALIVE) {
+    uint32_t activeLength = ent->mActiveConns.Length();
+    for (uint32_t i = 0; i < activeLength; i++) {
+      nsAHttpTransaction* activeTrans = ent->mActiveConns[i]->Transaction();
+      NullHttpTransaction* nullTrans =
+          activeTrans ? activeTrans->QueryNullTransaction() : nullptr;
+      if (nullTrans && nullTrans->Claim()) {
+        LOG(
+            ("nsHttpConnectionMgr::MakeNewConnection [ci = %s] "
+             "Claiming a null transaction for later use\n",
+             ent->mConnInfo->HashKey().get()));
+        pendingTransInfo->mActiveConn = do_GetWeakReference(
+            static_cast<nsISupportsWeakReference*>(ent->mActiveConns[i]));
+        return NS_OK;
+      }
+    }
+  }
 
   // If this host is trying to negotiate a SPDY session right now,
   // don't create any new connections until the result of the
   // negotiation is known.
   if (!(trans->Caps() & NS_HTTP_DISALLOW_SPDY) &&
-      (trans->Caps() & NS_HTTP_ALLOW_KEEPALIVE) && ent->RestrictConnections()) {
+      (trans->Caps() & NS_HTTP_ALLOW_KEEPALIVE) && RestrictConnections(ent)) {
     LOG(
         ("nsHttpConnectionMgr::MakeNewConnection [ci = %s] "
          "Not Available Due to RestrictConnections()\n",
@@ -1198,10 +1509,16 @@ nsresult nsHttpConnectionMgr::MakeNewConnection(
     // regardless of their TTL.
     auto iter = mCT.Iter();
     while (mNumIdleConns + mNumActiveConns + 1 >= mMaxConns && !iter.Done()) {
-      RefPtr<ConnectionEntry> entry = iter.Data();
-      entry->CloseIdleConnections((mNumIdleConns + mNumActiveConns + 1) -
-                                  mMaxConns);
-      iter.Next();
+      RefPtr<nsConnectionEntry> entry = iter.Data();
+      if (!entry->mIdleConns.Length()) {
+        iter.Next();
+        continue;
+      }
+      RefPtr<nsHttpConnection> conn(entry->mIdleConns[0]);
+      entry->mIdleConns.RemoveElementAt(0);
+      conn->Close(NS_ERROR_ABORT);
+      mNumIdleConns--;
+      ConditionallyStopPruneDeadConnectionsTimer();
     }
   }
 
@@ -1211,12 +1528,20 @@ nsresult nsHttpConnectionMgr::MakeNewConnection(
     // connections to a host without idle connections, then close any spdy
     // ASAP.
     for (auto iter = mCT.Iter(); !iter.Done(); iter.Next()) {
-      RefPtr<ConnectionEntry> entry = iter.Data();
-      while (entry->MakeFirstActiveSpdyConnDontReuse()) {
-        // Stop on <= (particularly =) because this dontreuse
-        // causes async close.
-        if (mNumIdleConns + mNumActiveConns + 1 <= mMaxConns) {
-          goto outerLoopEnd;
+      RefPtr<nsConnectionEntry> entry = iter.Data();
+      if (!entry->mUsingSpdy) {
+        continue;
+      }
+
+      for (uint32_t index = 0; index < entry->mActiveConns.Length(); ++index) {
+        HttpConnectionBase* conn = entry->mActiveConns[index];
+        if (conn->UsingSpdy() && conn->CanReuse()) {
+          conn->DontReuse();
+          // Stop on <= (particularly =) because this dontreuse
+          // causes async close.
+          if (mNumIdleConns + mNumActiveConns + 1 <= mMaxConns) {
+            goto outerLoopEnd;
+          }
         }
       }
     }
@@ -1251,19 +1576,21 @@ nsresult nsHttpConnectionMgr::MakeNewConnection(
 // returns other ERROR when transaction has a hard failure and should
 //   not remain in the pending queue
 nsresult nsHttpConnectionMgr::TryDispatchTransaction(
-    ConnectionEntry* ent, bool onlyReusedConnection,
+    nsConnectionEntry* ent, bool onlyReusedConnection,
     PendingTransactionInfo* pendingTransInfo) {
   MOZ_ASSERT(OnSocketThread(), "not on socket thread");
 
-  nsHttpTransaction* trans = pendingTransInfo->Transaction();
+  nsHttpTransaction* trans = pendingTransInfo->mTransaction;
 
   LOG(
       ("nsHttpConnectionMgr::TryDispatchTransaction without conn "
-       "[trans=%p ci=%p ci=%s caps=%x tunnelprovider=%p "
+       "[trans=%p halfOpen=%p conn=%p ci=%p ci=%s caps=%x tunnelprovider=%p "
        "onlyreused=%d active=%zu idle=%zu]\n",
-       trans, ent->mConnInfo.get(), ent->mConnInfo->HashKey().get(),
-       uint32_t(trans->Caps()), trans->TunnelProvider(), onlyReusedConnection,
-       ent->ActiveConnsLength(), ent->IdleConnectionsLength()));
+       trans, pendingTransInfo->mHalfOpen.get(),
+       pendingTransInfo->mActiveConn.get(), ent->mConnInfo.get(),
+       ent->mConnInfo->HashKey().get(), uint32_t(trans->Caps()),
+       trans->TunnelProvider(), onlyReusedConnection,
+       ent->mActiveConns.Length(), ent->mIdleConns.Length()));
 
   uint32_t caps = trans->Caps();
 
@@ -1412,7 +1739,7 @@ nsresult nsHttpConnectionMgr::TryDispatchTransaction(
     // dispatch deadlocks
     if (!(trans->ClassOfService() & nsIClassOfService::UrgentStart) &&
         idleConnsAllUrgent &&
-        ent->ActiveConnsLength() < MaxPersistConnections(ent)) {
+        ent->mActiveConns.Length() < MaxPersistConnections(ent)) {
       rv = TryDispatchTransactionOnIdleConn(ent, pendingTransInfo, false);
       if (NS_SUCCEEDED(rv)) {
         LOG(("   dispatched step 2a (idle, reuse urgent) trans=%p\n", trans));
@@ -1443,11 +1770,11 @@ nsresult nsHttpConnectionMgr::TryDispatchTransaction(
 }
 
 nsresult nsHttpConnectionMgr::TryDispatchTransactionOnIdleConn(
-    ConnectionEntry* ent, PendingTransactionInfo* pendingTransInfo,
+    nsConnectionEntry* ent, PendingTransactionInfo* pendingTransInfo,
     bool respectUrgency, bool* allUrgent) {
-  bool onlyUrgent = !!ent->IdleConnectionsLength();
+  bool onlyUrgent = !!ent->mIdleConns.Length();
 
-  nsHttpTransaction* trans = pendingTransInfo->Transaction();
+  nsHttpTransaction* trans = pendingTransInfo->mTransaction;
   bool urgentTrans = trans->ClassOfService() & nsIClassOfService::UrgentStart;
 
   LOG(
@@ -1455,8 +1782,40 @@ nsresult nsHttpConnectionMgr::TryDispatchTransactionOnIdleConn(
        "trans=%p, urgent=%d",
        ent, trans, urgentTrans));
 
-  RefPtr<nsHttpConnection> conn =
-      ent->GetIdleConnection(respectUrgency, urgentTrans, &onlyUrgent);
+  RefPtr<nsHttpConnection> conn;
+  size_t index = 0;
+  while (!conn && (ent->mIdleConns.Length() > index)) {
+    conn = ent->mIdleConns[index];
+
+    // non-urgent transactions can only be dispatched on non-urgent
+    // started or used connections.
+    if (respectUrgency && conn->IsUrgentStartPreferred() && !urgentTrans) {
+      LOG(("  skipping urgent: [conn=%p]", conn.get()));
+      conn = nullptr;
+      ++index;
+      continue;
+    }
+
+    onlyUrgent = false;
+
+    ent->mIdleConns.RemoveElementAt(index);
+    mNumIdleConns--;
+
+    // we check if the connection can be reused before even checking if
+    // it is a "matching" connection.
+    if (!conn->CanReuse()) {
+      LOG(("   dropping stale connection: [conn=%p]\n", conn.get()));
+      conn->Close(NS_ERROR_ABORT);
+      conn = nullptr;
+    } else {
+      LOG(("   reusing connection: [conn=%p]\n", conn.get()));
+      conn->EndIdleMonitoring();
+    }
+
+    // If there are no idle connections left at all, we need to make
+    // sure that we are not pruning dead connections anymore.
+    ConditionallyStopPruneDeadConnectionsTimer();
+  }
 
   if (allUrgent) {
     *allUrgent = onlyUrgent;
@@ -1465,7 +1824,7 @@ nsresult nsHttpConnectionMgr::TryDispatchTransactionOnIdleConn(
   if (conn) {
     // This will update the class of the connection to be the class of
     // the transaction dispatched on it.
-    ent->InsertIntoActiveConns(conn);
+    AddActiveConn(conn, ent);
     nsresult rv = DispatchTransaction(ent, trans, conn);
     NS_ENSURE_SUCCESS(rv, rv);
 
@@ -1475,7 +1834,7 @@ nsresult nsHttpConnectionMgr::TryDispatchTransactionOnIdleConn(
   return NS_ERROR_NOT_AVAILABLE;
 }
 
-nsresult nsHttpConnectionMgr::DispatchTransaction(ConnectionEntry* ent,
+nsresult nsHttpConnectionMgr::DispatchTransaction(nsConnectionEntry* ent,
                                                   nsHttpTransaction* trans,
                                                   HttpConnectionBase* conn) {
   uint32_t caps = trans->Caps();
@@ -1527,17 +1886,53 @@ nsresult nsHttpConnectionMgr::DispatchTransaction(ConnectionEntry* ent,
   return rv;
 }
 
+//-----------------------------------------------------------------------------
+// ConnectionHandle
+//
+// thin wrapper around a real connection, used to keep track of references
+// to the connection to determine when the connection may be reused.  the
+// transaction owns a reference to this handle.  this extra
+// layer of indirection greatly simplifies consumer code, avoiding the
+// need for consumer code to know when to give the connection back to the
+// connection manager.
+//
+class ConnectionHandle : public nsAHttpConnection {
+ public:
+  NS_DECL_THREADSAFE_ISUPPORTS
+  NS_DECL_NSAHTTPCONNECTION(mConn)
+
+  explicit ConnectionHandle(HttpConnectionBase* conn) : mConn(conn) {}
+  void Reset() { mConn = nullptr; }
+
+ private:
+  virtual ~ConnectionHandle();
+  RefPtr<HttpConnectionBase> mConn;
+};
+
 nsAHttpConnection* nsHttpConnectionMgr::MakeConnectionHandle(
     HttpConnectionBase* aWrapped) {
   return new ConnectionHandle(aWrapped);
 }
+
+ConnectionHandle::~ConnectionHandle() {
+  if (mConn) {
+    nsresult rv = gHttpHandler->ReclaimConnection(mConn);
+    if (NS_FAILED(rv)) {
+      LOG(
+          ("ConnectionHandle::~ConnectionHandle\n"
+           "    failed to reclaim connection\n"));
+    }
+  }
+}
+
+NS_IMPL_ISUPPORTS0(ConnectionHandle)
 
 // Use this method for dispatching nsAHttpTransction's. It can only safely be
 // used upon first use of a connection when NPN has not negotiated SPDY vs
 // HTTP/1 yet as multiplexing onto an existing SPDY session requires a
 // concrete nsHttpTransaction
 nsresult nsHttpConnectionMgr::DispatchAbstractTransaction(
-    ConnectionEntry* ent, nsAHttpTransaction* aTrans, uint32_t caps,
+    nsConnectionEntry* ent, nsAHttpTransaction* aTrans, uint32_t caps,
     HttpConnectionBase* conn, int32_t priority) {
   MOZ_ASSERT(ent);
 
@@ -1559,8 +1954,9 @@ nsresult nsHttpConnectionMgr::DispatchAbstractTransaction(
   if (NS_FAILED(rv)) {
     LOG(("  conn->Activate failed [rv=%" PRIx32 "]\n",
          static_cast<uint32_t>(rv)));
-    DebugOnly<nsresult> rv_remove = ent->RemoveActiveConnection(conn);
-    MOZ_ASSERT(NS_SUCCEEDED(rv_remove));
+    ent->mActiveConns.RemoveElement(conn);
+    DecrementActiveConnCount(conn);
+    ConditionallyStopTimeoutTick();
 
     // sever back references to connection, and do so without triggering
     // a call to ReclaimConnection ;-)
@@ -1571,7 +1967,7 @@ nsresult nsHttpConnectionMgr::DispatchAbstractTransaction(
   return rv;
 }
 
-void nsHttpConnectionMgr::ReportProxyTelemetry(ConnectionEntry* ent) {
+void nsHttpConnectionMgr::ReportProxyTelemetry(nsConnectionEntry* ent) {
   enum { PROXY_NONE = 1, PROXY_HTTP = 2, PROXY_SOCKS = 3, PROXY_HTTPS = 4 };
 
   if (!ent->mConnInfo->UsingProxy())
@@ -1616,7 +2012,7 @@ nsresult nsHttpConnectionMgr::ProcessNewTransaction(nsHttpTransaction* trans) {
   nsHttpConnectionInfo* ci = trans->ConnectionInfo();
   MOZ_ASSERT(ci);
 
-  ConnectionEntry* ent = GetOrCreateConnectionEntry(
+  nsConnectionEntry* ent = GetOrCreateConnectionEntry(
       ci, !!trans->TunnelProvider(), trans->Caps() & NS_HTTP_DISALLOW_HTTP3);
   MOZ_ASSERT(ent);
 
@@ -1638,7 +2034,7 @@ nsresult nsHttpConnectionMgr::ProcessNewTransaction(nsHttpTransaction* trans) {
          "sticky connection=%p\n",
          trans, conn.get()));
 
-    if (!ent->IsInActiveConns(conn)) {
+    if (static_cast<int32_t>(ent->mActiveConns.IndexOf(conn)) == -1) {
       LOG(
           ("nsHttpConnectionMgr::ProcessNewTransaction trans=%p "
            "sticky connection=%p needs to go on the active list\n",
@@ -1646,11 +2042,10 @@ nsresult nsHttpConnectionMgr::ProcessNewTransaction(nsHttpTransaction* trans) {
 
       // make sure it isn't on the idle list - we expect this to be an
       // unknown fresh connection
-      RefPtr<nsHttpConnection> connTCP = do_QueryObject(conn);
-      MOZ_ASSERT(!connTCP || !ent->IsInIdleConnections(connTCP));
+      MOZ_ASSERT(static_cast<int32_t>(ent->mIdleConns.IndexOf(conn)) == -1);
       MOZ_ASSERT(!conn->IsExperienced());
 
-      ent->InsertIntoActiveConns(conn);  // make it active
+      AddActiveConn(conn, ent);  // make it active
     }
 
     trans->SetConnection(nullptr);
@@ -1673,8 +2068,21 @@ nsresult nsHttpConnectionMgr::ProcessNewTransaction(nsHttpTransaction* trans) {
     if (!pendingTransInfo) {
       pendingTransInfo = new PendingTransactionInfo(trans);
     }
-
-    ent->InsertTransaction(pendingTransInfo);
+    if (trans->Caps() & NS_HTTP_URGENT_START) {
+      LOG(
+          ("  adding transaction to pending queue "
+           "[trans=%p urgent-start-count=%zu]\n",
+           trans, ent->mUrgentStartQ.Length() + 1));
+      // put this transaction on the urgent-start queue...
+      InsertTransactionSorted(ent->mUrgentStartQ, pendingTransInfo);
+    } else {
+      LOG(
+          ("  adding transaction to pending queue "
+           "[trans=%p pending-count=%zu]\n",
+           trans, ent->PendingQLength() + 1));
+      // put this transaction on the pending queue...
+      ent->InsertTransaction(pendingTransInfo);
+    }
     return NS_OK;
   }
 
@@ -1683,7 +2091,9 @@ nsresult nsHttpConnectionMgr::ProcessNewTransaction(nsHttpTransaction* trans) {
   return rv;
 }
 
-void nsHttpConnectionMgr::IncrementActiveConnCount() {
+void nsHttpConnectionMgr::AddActiveConn(HttpConnectionBase* conn,
+                                        nsConnectionEntry* ent) {
+  ent->mActiveConns.AppendElement(conn);
   mNumActiveConns++;
   ActivateTimeoutTick();
 }
@@ -1693,7 +2103,6 @@ void nsHttpConnectionMgr::DecrementActiveConnCount(HttpConnectionBase* conn) {
 
   RefPtr<nsHttpConnection> connTCP = do_QueryObject(conn);
   if (!connTCP || connTCP->EverUsedSpdy()) mNumSpdyHttp3ActiveConns--;
-  ConditionallyStopTimeoutTick();
 }
 
 void nsHttpConnectionMgr::StartedConnect() {
@@ -1706,15 +2115,42 @@ void nsHttpConnectionMgr::RecvdConnect() {
   ConditionallyStopTimeoutTick();
 }
 
+void nsHttpConnectionMgr::ReleaseClaimedSockets(
+    nsConnectionEntry* ent, PendingTransactionInfo* pendingTransInfo) {
+  if (pendingTransInfo->mHalfOpen) {
+    RefPtr<nsHalfOpenSocket> halfOpen =
+        do_QueryReferent(pendingTransInfo->mHalfOpen);
+    LOG(
+        ("nsHttpConnectionMgr::ReleaseClaimedSockets "
+         "[trans=%p halfOpen=%p]",
+         pendingTransInfo->mTransaction.get(), halfOpen.get()));
+    if (halfOpen) {
+      halfOpen->Unclaim();
+    }
+    pendingTransInfo->mHalfOpen = nullptr;
+  } else if (pendingTransInfo->mActiveConn) {
+    RefPtr<HttpConnectionBase> activeConn =
+        do_QueryReferent(pendingTransInfo->mActiveConn);
+    if (activeConn && activeConn->Transaction() &&
+        activeConn->Transaction()->IsNullTransaction()) {
+      NullHttpTransaction* nullTrans =
+          activeConn->Transaction()->QueryNullTransaction();
+      nullTrans->Unclaim();
+      LOG(("nsHttpConnectionMgr::ReleaseClaimedSockets - mark %p unclaimed.",
+           activeConn.get()));
+    }
+  }
+}
+
 nsresult nsHttpConnectionMgr::CreateTransport(
-    ConnectionEntry* ent, nsAHttpTransaction* trans, uint32_t caps,
+    nsConnectionEntry* ent, nsAHttpTransaction* trans, uint32_t caps,
     bool speculative, bool isFromPredictor, bool urgentStart, bool allow1918,
     PendingTransactionInfo* pendingTransInfo) {
   MOZ_ASSERT(OnSocketThread(), "not on socket thread");
   MOZ_ASSERT((speculative && !pendingTransInfo) ||
              (!speculative && pendingTransInfo));
 
-  RefPtr<HalfOpenSocket> sock = new HalfOpenSocket(
+  RefPtr<nsHalfOpenSocket> sock = new nsHalfOpenSocket(
       ent, trans, caps, speculative, isFromPredictor, urgentStart);
 
   if (speculative) {
@@ -1727,16 +2163,19 @@ nsresult nsHttpConnectionMgr::CreateTransport(
   NS_ENSURE_SUCCESS(rv, rv);
 
   if (pendingTransInfo) {
-    DebugOnly<bool> claimed = pendingTransInfo->TryClaimingHalfOpen(sock);
+    pendingTransInfo->mHalfOpen =
+        do_GetWeakReference(static_cast<nsISupportsWeakReference*>(sock));
+    DebugOnly<bool> claimed = sock->Claim();
     MOZ_ASSERT(claimed);
   }
 
-  ent->InsertIntoHalfOpens(sock);
+  ent->mHalfOpens.AppendElement(sock);
+  mNumHalfOpenConns++;
   return NS_OK;
 }
 
 void nsHttpConnectionMgr::DispatchSpdyPendingQ(
-    nsTArray<RefPtr<PendingTransactionInfo>>& pendingQ, ConnectionEntry* ent,
+    nsTArray<RefPtr<PendingTransactionInfo>>& pendingQ, nsConnectionEntry* ent,
     HttpConnectionBase* conn) {
   if (pendingQ.Length() == 0) {
     return;
@@ -1749,22 +2188,23 @@ void nsHttpConnectionMgr::DispatchSpdyPendingQ(
        ++index) {
     PendingTransactionInfo* pendingTransInfo = pendingQ[index];
 
-    if (!(pendingTransInfo->Transaction()->Caps() & NS_HTTP_ALLOW_KEEPALIVE) ||
-        pendingTransInfo->Transaction()->Caps() & NS_HTTP_DISALLOW_SPDY) {
+    if (!(pendingTransInfo->mTransaction->Caps() & NS_HTTP_ALLOW_KEEPALIVE) ||
+        pendingTransInfo->mTransaction->Caps() & NS_HTTP_DISALLOW_SPDY) {
       leftovers.AppendElement(pendingTransInfo);
       continue;
     }
 
     nsresult rv =
-        DispatchTransaction(ent, pendingTransInfo->Transaction(), conn);
+        DispatchTransaction(ent, pendingTransInfo->mTransaction, conn);
     if (NS_FAILED(rv)) {
       // this cannot happen, but if due to some bug it does then
       // close the transaction
       MOZ_ASSERT(false, "Dispatch SPDY Transaction");
       LOG(("ProcessSpdyPendingQ Dispatch Transaction failed trans=%p\n",
-           pendingTransInfo->Transaction()));
-      pendingTransInfo->Transaction()->Close(rv);
+           pendingTransInfo->mTransaction.get()));
+      pendingTransInfo->mTransaction->Close(rv);
     }
+    ReleaseClaimedSockets(ent, pendingTransInfo);
   }
 
   // Slurp up the rest of the pending queue into our leftovers bucket (we
@@ -1783,18 +2223,13 @@ void nsHttpConnectionMgr::DispatchSpdyPendingQ(
 // the connection entry sent in as an argument. It will do so on the
 // active h2 or h3 connection either in that same entry or from the
 // coalescing hash table
-void nsHttpConnectionMgr::ProcessSpdyPendingQ(ConnectionEntry* ent) {
+void nsHttpConnectionMgr::ProcessSpdyPendingQ(nsConnectionEntry* ent) {
   HttpConnectionBase* conn = GetH2orH3ActiveConn(ent, false);
   if (!conn || !conn->CanDirectlyActivate()) {
     return;
   }
 
-  nsTArray<RefPtr<PendingTransactionInfo>> urgentQ;
-  ent->AppendPendingUrgentStartQ(urgentQ);
-  DispatchSpdyPendingQ(urgentQ, ent, conn);
-  for (const auto& transactionInfo : Reversed(urgentQ)) {
-    ent->InsertTransaction(transactionInfo);
-  }
+  DispatchSpdyPendingQ(ent->mUrgentStartQ, ent, conn);
   if (!conn->CanDirectlyActivate()) {
     return;
   }
@@ -1821,16 +2256,65 @@ void nsHttpConnectionMgr::OnMsgProcessAllSpdyPendingQ(int32_t, ARefBase*) {
 // Given a connection entry, return an active h2 or h3 connection
 // that can be directly activated or null.
 HttpConnectionBase* nsHttpConnectionMgr::GetH2orH3ActiveConn(
-    ConnectionEntry* ent, bool aNoHttp3) {
+    nsConnectionEntry* ent, bool aNoHttp3) {
   MOZ_ASSERT(OnSocketThread(), "not on socket thread");
   MOZ_ASSERT(ent);
 
-  HttpConnectionBase* conn = ent->GetH2orH3ActiveConn(aNoHttp3);
-  if (conn) {
-    return conn;
+  HttpConnectionBase* experienced = nullptr;
+  HttpConnectionBase* noExperience = nullptr;
+  uint32_t activeLen = ent->mActiveConns.Length();
+  nsHttpConnectionInfo* ci = ent->mConnInfo;
+  uint32_t index;
+
+  // activeLen should generally be 1.. this is a setup race being resolved
+  // take a conn who can activate and is experienced
+  for (index = 0; index < activeLen; ++index) {
+    HttpConnectionBase* tmp = ent->mActiveConns[index];
+    if (tmp->CanDirectlyActivate()) {
+      if (tmp->IsExperienced()) {
+        experienced = tmp;
+        break;
+      }
+      noExperience = tmp;  // keep looking for a better option
+    }
   }
 
-  nsHttpConnectionInfo* ci = ent->mConnInfo;
+  // if that worked, cleanup anything else and exit
+  if (experienced) {
+    for (index = 0; index < activeLen; ++index) {
+      HttpConnectionBase* tmp = ent->mActiveConns[index];
+      // in the case where there is a functional h2 session, drop the others
+      if (tmp != experienced) {
+        tmp->DontReuse();
+      }
+    }
+    for (int32_t index = ent->mHalfOpenFastOpenBackups.Length() - 1; index >= 0;
+         --index) {
+      LOG(
+          ("GetH2orH3ActiveConn() shutting down connection in fast "
+           "open state (%p) because we have an experienced spdy "
+           "connection (%p).\n",
+           ent->mHalfOpenFastOpenBackups[index].get(), experienced));
+      RefPtr<nsHalfOpenSocket> half = ent->mHalfOpenFastOpenBackups[index];
+      half->CancelFastOpenConnection();
+    }
+
+    LOG(
+        ("GetH2orH3ActiveConn() request for ent %p %s "
+         "found an active experienced connection %p in native connection "
+         "entry\n",
+         ent, ci->HashKey().get(), experienced));
+    return experienced;
+  }
+
+  if (noExperience) {
+    LOG(
+        ("GetH2orH3ActiveConn() request for ent %p %s "
+         "found an active but inexperienced connection %p in native connection "
+         "entry\n",
+         ent, ci->HashKey().get(), noExperience));
+    return noExperience;
+  }
 
   // there was no active spdy connection in the connection entry, but
   // there might be one in the hash table for coalescing
@@ -1862,21 +2346,57 @@ void nsHttpConnectionMgr::AbortAndCloseAllConnections(int32_t, ARefBase*) {
   MOZ_ASSERT(OnSocketThread(), "not on socket thread");
   LOG(("nsHttpConnectionMgr::AbortAndCloseAllConnections\n"));
   for (auto iter = mCT.Iter(); !iter.Done(); iter.Next()) {
-    RefPtr<ConnectionEntry> ent = iter.Data();
+    RefPtr<nsConnectionEntry> ent = iter.Data();
 
     // Close all active connections.
-    ent->CloseActiveConnections();
+    while (ent->mActiveConns.Length()) {
+      RefPtr<HttpConnectionBase> conn(ent->mActiveConns[0]);
+      ent->mActiveConns.RemoveElementAt(0);
+      DecrementActiveConnCount(conn);
+      // Since HttpConnectionBase::Close doesn't break the bond with
+      // the connection's transaction, we must explicitely tell it
+      // to close its transaction and not just self.
+      conn->CloseTransaction(conn->Transaction(), NS_ERROR_ABORT, true);
+    }
 
     // Close all idle connections.
-    ent->CloseIdleConnections();
+    while (ent->mIdleConns.Length()) {
+      RefPtr<nsHttpConnection> conn(ent->mIdleConns[0]);
+
+      ent->mIdleConns.RemoveElementAt(0);
+      mNumIdleConns--;
+
+      conn->Close(NS_ERROR_ABORT);
+    }
+
+    // If all idle connections are removed we can stop pruning dead
+    // connections.
+    ConditionallyStopPruneDeadConnectionsTimer();
+
+    // Close all urgentStart transactions.
+    while (ent->mUrgentStartQ.Length()) {
+      PendingTransactionInfo* pendingTransInfo = ent->mUrgentStartQ[0];
+      pendingTransInfo->mTransaction->Close(NS_ERROR_ABORT);
+      ent->mUrgentStartQ.RemoveElementAt(0);
+    }
 
     // Close all pending transactions.
-    ent->CancelAllTransactions(NS_ERROR_ABORT);
+    for (auto it = ent->mPendingTransactionTable.Iter(); !it.Done();
+         it.Next()) {
+      while (it.UserData()->Length()) {
+        PendingTransactionInfo* pendingTransInfo = (*it.UserData())[0];
+        pendingTransInfo->mTransaction->Close(NS_ERROR_ABORT);
+        it.UserData()->RemoveElementAt(0);
+      }
+    }
+    ent->mPendingTransactionTable.Clear();
 
     // Close all half open tcp connections.
-    ent->CloseAllHalfOpens();
+    for (int32_t i = int32_t(ent->mHalfOpens.Length()) - 1; i >= 0; i--) {
+      ent->mHalfOpens[i]->Abandon();
+    }
 
-    MOZ_ASSERT(ent->HalfOpenFastOpenBackupsLength() == 0 &&
+    MOZ_ASSERT(ent->mHalfOpenFastOpenBackups.Length() == 0 &&
                !ent->mDoNotDestroy);
     iter.Remove();
   }
@@ -1970,6 +2490,25 @@ void nsHttpConnectionMgr::OnMsgNewTransactionWithStickyConn(int32_t priority,
   }
 }
 
+static uint64_t TabIdForQueuing(nsAHttpTransaction* transaction) {
+  return gHttpHandler->ActiveTabPriority()
+             ? transaction->TopLevelOuterContentWindowId()
+             : 0;
+}
+
+nsTArray<RefPtr<nsHttpConnectionMgr::PendingTransactionInfo>>*
+nsHttpConnectionMgr::GetTransactionPendingQHelper(nsConnectionEntry* ent,
+                                                  nsAHttpTransaction* trans) {
+  nsTArray<RefPtr<PendingTransactionInfo>>* pendingQ = nullptr;
+  int32_t caps = trans->Caps();
+  if (caps & NS_HTTP_URGENT_START) {
+    pendingQ = &(ent->mUrgentStartQ);
+  } else {
+    pendingQ = ent->mPendingTransactionTable.Get(TabIdForQueuing(trans));
+  }
+  return pendingQ;
+}
+
 void nsHttpConnectionMgr::OnMsgReschedTransaction(int32_t priority,
                                                   ARefBase* param) {
   MOZ_ASSERT(OnSocketThread(), "not on socket thread");
@@ -1981,10 +2520,19 @@ void nsHttpConnectionMgr::OnMsgReschedTransaction(int32_t priority,
   if (!trans->ConnectionInfo()) {
     return;
   }
-  ConnectionEntry* ent = mCT.GetWeak(trans->ConnectionInfo()->HashKey());
+  nsConnectionEntry* ent = mCT.GetWeak(trans->ConnectionInfo()->HashKey());
 
   if (ent) {
-    ent->ReschedTransaction(trans);
+    nsTArray<RefPtr<PendingTransactionInfo>>* pendingQ =
+        GetTransactionPendingQHelper(ent, trans);
+
+    int32_t index =
+        pendingQ ? pendingQ->IndexOf(trans, 0, PendingComparator()) : -1;
+    if (index >= 0) {
+      RefPtr<PendingTransactionInfo> pendingTransInfo = (*pendingQ)[index];
+      pendingQ->RemoveElementAt(index);
+      InsertTransactionSorted(*pendingQ, pendingTransInfo);
+    }
   }
 }
 
@@ -2026,7 +2574,7 @@ void nsHttpConnectionMgr::OnMsgCancelTransaction(int32_t reason,
   if (conn && !trans->IsDone()) {
     conn->CloseTransaction(trans, closeCode);
   } else {
-    ConnectionEntry* ent = nullptr;
+    nsConnectionEntry* ent = nullptr;
     if (trans->ConnectionInfo()) {
       ent = mCT.GetWeak(trans->ConnectionInfo()->HashKey());
     }
@@ -2035,7 +2583,7 @@ void nsHttpConnectionMgr::OnMsgCancelTransaction(int32_t reason,
       // We will abandon all half-open sockets belonging to the given
       // transaction.
       nsTArray<RefPtr<PendingTransactionInfo>>* infoArray =
-          ent->GetTransactionPendingQHelper(trans);
+          GetTransactionPendingQHelper(ent, trans);
 
       RefPtr<PendingTransactionInfo> pendingTransInfo;
       transIndex =
@@ -2046,12 +2594,19 @@ void nsHttpConnectionMgr::OnMsgCancelTransaction(int32_t reason,
              " found in urgentStart queue\n",
              trans));
         pendingTransInfo = (*infoArray)[transIndex];
+        // We do not need to ReleaseClaimedSockets while we are
+        // going to close them all any way!
         infoArray->RemoveElementAt(transIndex);
       }
 
       // Abandon all half-open sockets belonging to the given transaction.
       if (pendingTransInfo) {
-        pendingTransInfo->AbandonHalfOpenAndForgetActiveConn();
+        RefPtr<nsHalfOpenSocket> half =
+            do_QueryReferent(pendingTransInfo->mHalfOpen);
+        if (half) {
+          half->Abandon();
+        }
+        pendingTransInfo->mHalfOpen = nullptr;
       }
     }
 
@@ -2063,8 +2618,17 @@ void nsHttpConnectionMgr::OnMsgCancelTransaction(int32_t reason,
     // anything that might get canceled from the rest of gecko, so best
     // to assume that's what was meant by the cancel we did receive if
     // it only applied to something in the queue.
-    if (ent) {
-      ent->CloseAllActiveConnsWithNullTransactcion(closeCode);
+    for (uint32_t index = 0; ent && (index < ent->mActiveConns.Length());
+         ++index) {
+      HttpConnectionBase* activeConn = ent->mActiveConns[index];
+      nsAHttpTransaction* liveTransaction = activeConn->Transaction();
+      if (liveTransaction && liveTransaction->IsNullTransaction()) {
+        LOG(
+            ("nsHttpConnectionMgr::OnMsgCancelTransaction [trans=%p] "
+             "also canceling Null Transaction %p on conn %p\n",
+             trans, liveTransaction, activeConn));
+        activeConn->CloseTransaction(liveTransaction, closeCode);
+      }
     }
   }
 }
@@ -2086,7 +2650,7 @@ void nsHttpConnectionMgr::OnMsgProcessPendingQ(int32_t, ARefBase* param) {
        ci->HashKey().get()));
 
   // start by processing the queue identified by the given connection info.
-  ConnectionEntry* ent = mCT.GetWeak(ci->HashKey());
+  nsConnectionEntry* ent = mCT.GetWeak(ci->HashKey());
   if (!(ent && ProcessPendingQForEntry(ent, false))) {
     // if we reach here, it means that we couldn't dispatch a transaction
     // for the specified connection info.  walk the connection table...
@@ -2107,16 +2671,35 @@ nsresult nsHttpConnectionMgr::CancelTransactions(nsHttpConnectionInfo* ci,
                    ci);
 }
 
+void nsHttpConnectionMgr::CancelTransactionsHelper(
+    nsTArray<RefPtr<nsHttpConnectionMgr::PendingTransactionInfo>>& pendingQ,
+    const nsHttpConnectionInfo* ci,
+    const nsHttpConnectionMgr::nsConnectionEntry* ent, nsresult reason) {
+  for (const auto& pendingTransInfo : pendingQ) {
+    LOG(("nsHttpConnectionMgr::OnMsgCancelTransactions %s %p %p\n",
+         ci->HashKey().get(), ent, pendingTransInfo->mTransaction.get()));
+    pendingTransInfo->mTransaction->Close(reason);
+  }
+  pendingQ.Clear();
+}
+
 void nsHttpConnectionMgr::OnMsgCancelTransactions(int32_t code,
                                                   ARefBase* param) {
   nsresult reason = static_cast<nsresult>(code);
   nsHttpConnectionInfo* ci = static_cast<nsHttpConnectionInfo*>(param);
-  ConnectionEntry* ent = mCT.GetWeak(ci->HashKey());
+  nsConnectionEntry* ent = mCT.GetWeak(ci->HashKey());
   LOG(("nsHttpConnectionMgr::OnMsgCancelTransactions %s %p\n",
        ci->HashKey().get(), ent));
-  if (ent) {
-    ent->CancelAllTransactions(reason);
+  if (!ent) {
+    return;
   }
+
+  CancelTransactionsHelper(ent->mUrgentStartQ, ci, ent, reason);
+
+  for (auto it = ent->mPendingTransactionTable.Iter(); !it.Done(); it.Next()) {
+    CancelTransactionsHelper(*it.UserData(), ci, ent, reason);
+  }
+  ent->mPendingTransactionTable.Clear();
 }
 
 void nsHttpConnectionMgr::OnMsgPruneDeadConnections(int32_t, ARefBase*) {
@@ -2130,13 +2713,44 @@ void nsHttpConnectionMgr::OnMsgPruneDeadConnections(int32_t, ARefBase*) {
   // connection entries that are using spdy.
   if (mNumIdleConns || (mNumActiveConns && gHttpHandler->IsSpdyEnabled())) {
     for (auto iter = mCT.Iter(); !iter.Done(); iter.Next()) {
-      RefPtr<ConnectionEntry> ent = iter.Data();
+      RefPtr<nsConnectionEntry> ent = iter.Data();
 
       LOG(("  pruning [ci=%s]\n", ent->mConnInfo->HashKey().get()));
 
       // Find out how long it will take for next idle connection to not
       // be reusable anymore.
-      uint32_t timeToNextExpire = ent->PruneDeadConnections();
+      uint32_t timeToNextExpire = UINT32_MAX;
+      int32_t count = ent->mIdleConns.Length();
+      if (count > 0) {
+        for (int32_t i = count - 1; i >= 0; --i) {
+          RefPtr<nsHttpConnection> conn(ent->mIdleConns[i]);
+          if (!conn->CanReuse()) {
+            ent->mIdleConns.RemoveElementAt(i);
+            conn->Close(NS_ERROR_ABORT);
+            mNumIdleConns--;
+          } else {
+            timeToNextExpire = std::min(timeToNextExpire, conn->TimeToLive());
+          }
+        }
+      }
+
+      if (ent->mUsingSpdy) {
+        for (uint32_t i = 0; i < ent->mActiveConns.Length(); ++i) {
+          RefPtr<nsHttpConnection> connTCP =
+              do_QueryObject(ent->mActiveConns[i]);
+          // Http3 has its own timers, it is not using this one.
+          if (connTCP && connTCP->UsingSpdy()) {
+            if (!connTCP->CanReuse()) {
+              // Marking it don't-reuse will create an active
+              // tear down if the spdy session is idle.
+              connTCP->DontReuse();
+            } else {
+              timeToNextExpire =
+                  std::min(timeToNextExpire, connTCP->TimeToLive());
+            }
+          }
+        }
+      }
 
       // If time to next expire found is shorter than time to next
       // wake-up, we need to change the time for next wake-up.
@@ -2158,11 +2772,10 @@ void nsHttpConnectionMgr::OnMsgPruneDeadConnections(int32_t, ARefBase*) {
 
       // If this entry is empty, we have too many entries busy then
       // we can clean it up and restart
-      if (mCT.Count() > 125 && ent->IdleConnectionsLength() == 0 &&
-          ent->ActiveConnsLength() == 0 && ent->HalfOpensLength() == 0 &&
-          ent->PendingQueueLength() == 0 &&
-          ent->UrgentStartQueueLength() == 0 &&
-          ent->HalfOpenFastOpenBackupsLength() == 0 && !ent->mDoNotDestroy &&
+      if (mCT.Count() > 125 && ent->mIdleConns.Length() == 0 &&
+          ent->mActiveConns.Length() == 0 && ent->mHalfOpens.Length() == 0 &&
+          ent->PendingQLength() == 0 && ent->mUrgentStartQ.Length() == 0 &&
+          ent->mHalfOpenFastOpenBackups.Length() == 0 && !ent->mDoNotDestroy &&
           (!ent->mUsingSpdy || mCT.Count() > 300)) {
         LOG(("    removing empty connection entry\n"));
         iter.Remove();
@@ -2170,7 +2783,14 @@ void nsHttpConnectionMgr::OnMsgPruneDeadConnections(int32_t, ARefBase*) {
       }
 
       // Otherwise use this opportunity to compact our arrays...
-      ent->Compact();
+      ent->mIdleConns.Compact();
+      ent->mActiveConns.Compact();
+      ent->mUrgentStartQ.Compact();
+
+      for (auto it = ent->mPendingTransactionTable.Iter(); !it.Done();
+           it.Next()) {
+        it.UserData()->Compact();
+      }
     }
   }
 }
@@ -2182,8 +2802,29 @@ void nsHttpConnectionMgr::OnMsgPruneNoTraffic(int32_t, ARefBase*) {
   // Prune connections without traffic
   for (auto iter = mCT.Iter(); !iter.Done(); iter.Next()) {
     // Close the connections with no registered traffic.
-    RefPtr<ConnectionEntry> ent = iter.Data();
-    ent->PruneNoTraffic();
+    RefPtr<nsConnectionEntry> ent = iter.Data();
+
+    LOG(("  pruning no traffic [ci=%s]\n", ent->mConnInfo->HashKey().get()));
+
+    if (!ent->mConnInfo->IsHttp3()) {
+      uint32_t numConns = ent->mActiveConns.Length();
+      if (numConns) {
+        // Walk the list backwards to allow us to remove entries easily.
+        for (int index = numConns - 1; index >= 0; index--) {
+          RefPtr<nsHttpConnection> conn =
+              do_QueryObject(ent->mActiveConns[index]);
+          if (conn && conn->NoTraffic()) {
+            ent->mActiveConns.RemoveElementAt(index);
+            DecrementActiveConnCount(conn);
+            conn->Close(NS_ERROR_ABORT);
+            LOG(
+                ("  closed active connection due to no traffic "
+                 "[conn=%p]\n",
+                 conn.get()));
+          }
+        }
+      }
+    }
   }
 
   mPruningNoTraffic = false;  // not pruning anymore
@@ -2201,7 +2842,24 @@ void nsHttpConnectionMgr::OnMsgVerifyTraffic(int32_t, ARefBase*) {
 
   // Mark connections for traffic verification
   for (auto iter = mCT.Iter(); !iter.Done(); iter.Next()) {
-    iter.Data()->VerifyTraffic();
+    RefPtr<nsConnectionEntry> ent = iter.Data();
+    if (!ent->mConnInfo->IsHttp3()) {
+      // Iterate over all active connections and check them.
+      for (uint32_t index = 0; index < ent->mActiveConns.Length(); ++index) {
+        RefPtr<nsHttpConnection> conn =
+            do_QueryObject(ent->mActiveConns[index]);
+        if (conn) {
+          conn->CheckForTraffic(true);
+        }
+      }
+      // Iterate the idle connections and unmark them for traffic checks.
+      for (uint32_t index = 0; index < ent->mIdleConns.Length(); ++index) {
+        RefPtr<nsHttpConnection> conn = do_QueryObject(ent->mIdleConns[index]);
+        if (conn) {
+          conn->CheckForTraffic(false);
+        }
+      }
+    }
   }
 
   // If the timer is already there. we just re-init it
@@ -2229,7 +2887,7 @@ void nsHttpConnectionMgr::OnMsgDoShiftReloadConnectionCleanup(int32_t,
   nsHttpConnectionInfo* ci = static_cast<nsHttpConnectionInfo*>(param);
 
   for (auto iter = mCT.Iter(); !iter.Done(); iter.Next()) {
-    iter.Data()->ClosePersistentConnections();
+    ClosePersistentConnections(iter.Data());
   }
 
   if (ci) ResetIPFamilyPreference(ci);
@@ -2245,9 +2903,9 @@ void nsHttpConnectionMgr::OnMsgReclaimConnection(HttpConnectionBase* conn) {
   //
 
   MOZ_ASSERT(conn);
-  ConnectionEntry* ent = conn->ConnectionInfo()
-                             ? mCT.GetWeak(conn->ConnectionInfo()->HashKey())
-                             : nullptr;
+  nsConnectionEntry* ent = conn->ConnectionInfo()
+                               ? mCT.GetWeak(conn->ConnectionInfo()->HashKey())
+                               : nullptr;
 
   if (!ent) {
     // this can happen if the connection is made outside of the
@@ -2289,7 +2947,9 @@ void nsHttpConnectionMgr::OnMsgReclaimConnection(HttpConnectionBase* conn) {
     conn->DontReuse();
   }
 
-  if (NS_SUCCEEDED(ent->RemoveActiveConnection(conn))) {
+  if (ent->mActiveConns.RemoveElement(conn)) {
+    DecrementActiveConnCount(conn);
+    ConditionallyStopTimeoutTick();
   } else if (!connTCP || connTCP->EverUsedSpdy()) {
     LOG(("HttpConnectionBase %p not found in its connection entry, try ^anon",
          conn));
@@ -2298,9 +2958,11 @@ void nsHttpConnectionMgr::OnMsgReclaimConnection(HttpConnectionBase* conn) {
     RefPtr<nsHttpConnectionInfo> anonInvertedCI(ci->Clone());
     anonInvertedCI->SetAnonymous(!ci->GetAnonymous());
 
-    ConnectionEntry* ent = mCT.GetWeak(anonInvertedCI->HashKey());
+    nsConnectionEntry* ent = mCT.GetWeak(anonInvertedCI->HashKey());
     if (ent) {
-      if (NS_SUCCEEDED(ent->RemoveActiveConnection(conn))) {
+      if (ent->mActiveConns.RemoveElement(conn)) {
+        DecrementActiveConnCount(conn);
+        ConditionallyStopTimeoutTick();
       } else {
         LOG(
             ("HttpConnectionBase %p could not be removed from its entry's "
@@ -2319,7 +2981,22 @@ void nsHttpConnectionMgr::OnMsgReclaimConnection(HttpConnectionBase* conn) {
     // The linear search is ok here because the number of idleconns
     // in a single entry is generally limited to a small number (i.e. 6)
 
-    ent->InsertIntoIdleConnections(connTCP);
+    uint32_t idx;
+    for (idx = 0; idx < ent->mIdleConns.Length(); idx++) {
+      nsHttpConnection* idleConn = ent->mIdleConns[idx];
+      if (idleConn->MaxBytesRead() < connTCP->MaxBytesRead()) break;
+    }
+
+    ent->mIdleConns.InsertElementAt(idx, connTCP);
+    mNumIdleConns++;
+    connTCP->BeginIdleMonitoring();
+
+    // If the added connection was first idle connection or has shortest
+    // time to live among the watched connections, pruning dead
+    // connections needs to be done when it can't be reused anymore.
+    uint32_t timeToLive = connTCP->TimeToLive();
+    if (!mTimer || NowInSeconds() + timeToLive < mTimeOfNextWakeUp)
+      PruneDeadConnectionsAfter(timeToLive);
   } else {
     LOG(("  connection cannot be reused; closing connection\n"));
     conn->Close(NS_ERROR_ABORT);
@@ -2446,6 +3123,21 @@ void nsHttpConnectionMgr::OnMsgUpdateParam(int32_t inParam, ARefBase*) {
     default:
       MOZ_ASSERT_UNREACHABLE("unexpected parameter name");
   }
+}
+
+// nsHttpConnectionMgr::nsConnectionEntry
+nsHttpConnectionMgr::nsConnectionEntry::~nsConnectionEntry() {
+  LOG(("nsConnectionEntry::~nsConnectionEntry this=%p", this));
+
+  MOZ_ASSERT(!mIdleConns.Length());
+  MOZ_ASSERT(!mActiveConns.Length());
+  MOZ_ASSERT(!mHalfOpens.Length());
+  MOZ_ASSERT(!mUrgentStartQ.Length());
+  MOZ_ASSERT(!PendingQLength());
+  MOZ_ASSERT(!mHalfOpenFastOpenBackups.Length());
+  MOZ_ASSERT(!mDoNotDestroy);
+
+  MOZ_COUNT_DTOR(nsConnectionEntry);
 }
 
 // Read Timeout Tick handlers
@@ -2831,14 +3523,16 @@ bool nsHttpConnectionMgr::ShouldThrottle(nsHttpTransaction* aTrans) {
 
 bool nsHttpConnectionMgr::IsConnEntryUnderPressure(
     nsHttpConnectionInfo* connInfo) {
-  ConnectionEntry* ent = mCT.GetWeak(connInfo->HashKey());
+  nsConnectionEntry* ent = mCT.GetWeak(connInfo->HashKey());
   if (!ent) {
     // No entry, no pressure.
     return false;
   }
 
-  return ent->PendingQueueLengthForWindow(
-             mCurrentTopLevelOuterContentWindowId) > 0;
+  nsTArray<RefPtr<PendingTransactionInfo>>* transactions =
+      ent->mPendingTransactionTable.Get(mCurrentTopLevelOuterContentWindowId);
+
+  return transactions && !transactions->IsEmpty();
 }
 
 bool nsHttpConnectionMgr::IsThrottleTickerNeeded() {
@@ -3168,10 +3862,65 @@ void nsHttpConnectionMgr::TimeoutTick() {
   mTimeoutTickNext = 3600;  // 1hr
 
   for (auto iter = mCT.Iter(); !iter.Done(); iter.Next()) {
-    RefPtr<ConnectionEntry> ent = iter.Data();
+    RefPtr<nsConnectionEntry> ent = iter.Data();
 
-    uint32_t timeoutTickNext = ent->TimeoutTick();
-    mTimeoutTickNext = std::min(mTimeoutTickNext, timeoutTickNext);
+    if (!ent->mConnInfo->IsHttp3()) {
+      LOG(
+          ("nsHttpConnectionMgr::TimeoutTick() this=%p host=%s "
+           "idle=%zu active=%zu"
+           " half-len=%zu pending=%zu"
+           " urgentStart pending=%zu\n",
+           this, ent->mConnInfo->Origin(), ent->mIdleConns.Length(),
+           ent->mActiveConns.Length(), ent->mHalfOpens.Length(),
+           ent->PendingQLength(), ent->mUrgentStartQ.Length()));
+
+      // First call the tick handler for each active connection.
+      PRIntervalTime tickTime = PR_IntervalNow();
+      for (uint32_t index = 0; index < ent->mActiveConns.Length(); ++index) {
+        RefPtr<nsHttpConnection> conn =
+            do_QueryObject(ent->mActiveConns[index]);
+        if (conn) {
+          uint32_t connNextTimeout = conn->ReadTimeoutTick(tickTime);
+          mTimeoutTickNext = std::min(mTimeoutTickNext, connNextTimeout);
+        }
+      }
+
+      // Now check for any stalled half open sockets.
+      if (ent->mHalfOpens.Length()) {
+        TimeStamp currentTime = TimeStamp::Now();
+        double maxConnectTime_ms = gHttpHandler->ConnectTimeout();
+
+        for (uint32_t index = ent->mHalfOpens.Length(); index > 0;) {
+          index--;
+
+          nsHalfOpenSocket* half = ent->mHalfOpens[index];
+          double delta = half->Duration(currentTime);
+          // If the socket has timed out, close it so the waiting
+          // transaction will get the proper signal.
+          if (delta > maxConnectTime_ms) {
+            LOG(("Force timeout of half open to %s after %.2fms.\n",
+                 ent->mConnInfo->HashKey().get(), delta));
+            if (half->SocketTransport()) {
+              half->SocketTransport()->Close(NS_ERROR_NET_TIMEOUT);
+            }
+            if (half->BackupTransport()) {
+              half->BackupTransport()->Close(NS_ERROR_NET_TIMEOUT);
+            }
+          }
+
+          // If this half open hangs around for 5 seconds after we've
+          // closed() it then just abandon the socket.
+          if (delta > maxConnectTime_ms + 5000) {
+            LOG(("Abandon half open to %s after %.2fms.\n",
+                 ent->mConnInfo->HashKey().get(), delta));
+            half->Abandon();
+          }
+        }
+      }
+      if (ent->mHalfOpens.Length()) {
+        mTimeoutTickNext = 1;
+      }
+    }
   }
 
   if (mTimeoutTick) {
@@ -3186,10 +3935,11 @@ void nsHttpConnectionMgr::TimeoutTick() {
 // 2] otherwise use an ent of wildcard(ci) than can be dispatched immediately
 // 3] otherwise create an ent that matches ci and make new conn on it
 
-ConnectionEntry* nsHttpConnectionMgr::GetOrCreateConnectionEntry(
+nsHttpConnectionMgr::nsConnectionEntry*
+nsHttpConnectionMgr::GetOrCreateConnectionEntry(
     nsHttpConnectionInfo* specificCI, bool prohibitWildCard, bool aNoHttp3) {
   // step 1
-  ConnectionEntry* specificEnt = mCT.GetWeak(specificCI->HashKey());
+  nsConnectionEntry* specificEnt = mCT.GetWeak(specificCI->HashKey());
   if (specificEnt && specificEnt->AvailableForDispatchNow()) {
     return specificEnt;
   }
@@ -3199,7 +3949,7 @@ ConnectionEntry* nsHttpConnectionMgr::GetOrCreateConnectionEntry(
   // with a client certificate.
   RefPtr<nsHttpConnectionInfo> anonInvertedCI(specificCI->Clone());
   anonInvertedCI->SetAnonymous(!specificCI->GetAnonymous());
-  ConnectionEntry* invertedEnt = mCT.GetWeak(anonInvertedCI->HashKey());
+  nsConnectionEntry* invertedEnt = mCT.GetWeak(anonInvertedCI->HashKey());
   if (invertedEnt) {
     HttpConnectionBase* h2orh3conn = GetH2orH3ActiveConn(invertedEnt, aNoHttp3);
     if (h2orh3conn && h2orh3conn->IsExperienced() &&
@@ -3223,7 +3973,7 @@ ConnectionEntry* nsHttpConnectionMgr::GetOrCreateConnectionEntry(
     DebugOnly<nsresult> rv =
         specificCI->CreateWildCard(getter_AddRefs(wildCardProxyCI));
     MOZ_ASSERT(NS_SUCCEEDED(rv));
-    ConnectionEntry* wildCardEnt = mCT.GetWeak(wildCardProxyCI->HashKey());
+    nsConnectionEntry* wildCardEnt = mCT.GetWeak(wildCardProxyCI->HashKey());
     if (wildCardEnt && wildCardEnt->AvailableForDispatchNow()) {
       return wildCardEnt;
     }
@@ -3232,10 +3982,28 @@ ConnectionEntry* nsHttpConnectionMgr::GetOrCreateConnectionEntry(
   // step 3
   if (!specificEnt) {
     RefPtr<nsHttpConnectionInfo> clone(specificCI->Clone());
-    specificEnt = new ConnectionEntry(clone);
+    specificEnt = new nsConnectionEntry(clone);
     mCT.Put(clone->HashKey(), RefPtr{specificEnt});
   }
   return specificEnt;
+}
+
+nsresult ConnectionHandle::OnHeadersAvailable(nsAHttpTransaction* trans,
+                                              nsHttpRequestHead* req,
+                                              nsHttpResponseHead* resp,
+                                              bool* reset) {
+  return mConn->OnHeadersAvailable(trans, req, resp, reset);
+}
+
+void ConnectionHandle::CloseTransaction(nsAHttpTransaction* trans,
+                                        nsresult reason) {
+  mConn->CloseTransaction(trans, reason);
+}
+
+nsresult ConnectionHandle::TakeTransport(nsISocketTransport** aTransport,
+                                         nsIAsyncInputStream** aInputStream,
+                                         nsIAsyncOutputStream** aOutputStream) {
+  return mConn->TakeTransport(aTransport, aInputStream, aOutputStream);
 }
 
 void nsHttpConnectionMgr::DoSpeculativeConnection(
@@ -3243,7 +4011,7 @@ void nsHttpConnectionMgr::DoSpeculativeConnection(
   MOZ_ASSERT(OnSocketThread(), "not on socket thread");
   MOZ_ASSERT(aTrans);
 
-  ConnectionEntry* ent = GetOrCreateConnectionEntry(
+  nsConnectionEntry* ent = GetOrCreateConnectionEntry(
       aTrans->ConnectionInfo(), false, aTrans->Caps() & NS_HTTP_DISALLOW_HTTP3);
 
   uint32_t parallelSpeculativeConnectLimit =
@@ -3258,9 +4026,9 @@ void nsHttpConnectionMgr::DoSpeculativeConnection(
   bool keepAlive = aTrans->Caps() & NS_HTTP_ALLOW_KEEPALIVE;
   if (mNumHalfOpenConns < parallelSpeculativeConnectLimit &&
       ((ignoreIdle &&
-        (ent->IdleConnectionsLength() < parallelSpeculativeConnectLimit)) ||
-       !ent->IdleConnectionsLength()) &&
-      !(keepAlive && ent->RestrictConnections()) &&
+        (ent->mIdleConns.Length() < parallelSpeculativeConnectLimit)) ||
+       !ent->mIdleConns.Length()) &&
+      !(keepAlive && RestrictConnections(ent)) &&
       !AtActiveConnectionLimit(ent, aTrans->Caps())) {
     if (aFetchHTTPSRR) {
       Unused << aTrans->FetchHTTPSRR();
@@ -3288,6 +4056,95 @@ void nsHttpConnectionMgr::OnMsgSpeculativeConnect(int32_t, ARefBase* param) {
   DoSpeculativeConnection(args->mTrans, args->mFetchHTTPSRR);
 }
 
+bool ConnectionHandle::IsPersistent() {
+  MOZ_ASSERT(OnSocketThread());
+  return mConn->IsPersistent();
+}
+
+bool ConnectionHandle::IsReused() {
+  MOZ_ASSERT(OnSocketThread());
+  return mConn->IsReused();
+}
+
+void ConnectionHandle::DontReuse() {
+  MOZ_ASSERT(OnSocketThread());
+  mConn->DontReuse();
+}
+
+nsresult ConnectionHandle::PushBack(const char* buf, uint32_t bufLen) {
+  return mConn->PushBack(buf, bufLen);
+}
+
+//////////////////////// nsHalfOpenSocket
+NS_IMPL_ADDREF(nsHttpConnectionMgr::nsHalfOpenSocket)
+NS_IMPL_RELEASE(nsHttpConnectionMgr::nsHalfOpenSocket)
+
+NS_INTERFACE_MAP_BEGIN(nsHttpConnectionMgr::nsHalfOpenSocket)
+  NS_INTERFACE_MAP_ENTRY(nsISupportsWeakReference)
+  NS_INTERFACE_MAP_ENTRY(nsIOutputStreamCallback)
+  NS_INTERFACE_MAP_ENTRY(nsITransportEventSink)
+  NS_INTERFACE_MAP_ENTRY(nsIInterfaceRequestor)
+  NS_INTERFACE_MAP_ENTRY(nsITimerCallback)
+  NS_INTERFACE_MAP_ENTRY(nsINamed)
+  // we have no macro that covers this case.
+  if (aIID.Equals(NS_GET_IID(nsHttpConnectionMgr::nsHalfOpenSocket))) {
+    AddRef();
+    *aInstancePtr = this;
+    return NS_OK;
+  } else
+NS_INTERFACE_MAP_END
+
+nsHttpConnectionMgr::nsHalfOpenSocket::nsHalfOpenSocket(
+    nsConnectionEntry* ent, nsAHttpTransaction* trans, uint32_t caps,
+    bool speculative, bool isFromPredictor, bool urgentStart)
+    : mTransaction(trans),
+      mDispatchedMTransaction(false),
+      mCaps(caps),
+      mSpeculative(speculative),
+      mUrgentStart(urgentStart),
+      mIsFromPredictor(isFromPredictor),
+      mAllow1918(true),
+      mHasConnected(false),
+      mPrimaryConnectedOK(false),
+      mBackupConnectedOK(false),
+      mBackupConnStatsSet(false),
+      mFreeToUse(true),
+      mPrimaryStreamStatus(NS_OK),
+      mFastOpenInProgress(false),
+      mEnt(ent) {
+  MOZ_ASSERT(ent && trans, "constructor with null arguments");
+  LOG(("Creating nsHalfOpenSocket [this=%p trans=%p ent=%s key=%s]\n", this,
+       trans, ent->mConnInfo->Origin(), ent->mConnInfo->HashKey().get()));
+
+  mIsHttp3 = mEnt->mConnInfo->IsHttp3();
+  if (speculative) {
+    Telemetry::AutoCounter<Telemetry::HTTPCONNMGR_TOTAL_SPECULATIVE_CONN>
+        totalSpeculativeConn;
+    ++totalSpeculativeConn;
+
+    if (isFromPredictor) {
+      Telemetry::AutoCounter<Telemetry::PREDICTOR_TOTAL_PRECONNECTS_CREATED>
+          totalPreconnectsCreated;
+      ++totalPreconnectsCreated;
+    }
+  }
+
+  if (mEnt->mConnInfo->FirstHopSSL()) {
+    mFastOpenStatus = TFO_UNKNOWN;
+  } else {
+    mFastOpenStatus = TFO_HTTP;
+  }
+  MOZ_ASSERT(mEnt);
+}
+
+nsHttpConnectionMgr::nsHalfOpenSocket::~nsHalfOpenSocket() {
+  MOZ_ASSERT(!mStreamOut);
+  MOZ_ASSERT(!mBackupStreamOut);
+  LOG(("Destroying nsHalfOpenSocket [this=%p]\n", this));
+
+  if (mEnt) mEnt->RemoveHalfOpen(this);
+}
+
 bool nsHttpConnectionMgr::BeConservativeIfProxied(nsIProxyInfo* proxy) {
   if (mBeConservativeForProxy) {
     // The pref says to be conservative for proxies.
@@ -3304,6 +4161,1067 @@ bool nsHttpConnectionMgr::BeConservativeIfProxied(nsIProxyInfo* proxy) {
   nsAutoCString proxyHost;
   proxy->GetHost(proxyHost);
   return proxyHost.IsEmpty();
+}
+
+nsresult nsHttpConnectionMgr::nsHalfOpenSocket::SetupStreams(
+    nsISocketTransport** transport, nsIAsyncInputStream** instream,
+    nsIAsyncOutputStream** outstream, bool isBackup) {
+  MOZ_ASSERT(OnSocketThread(), "not on socket thread");
+
+  MOZ_ASSERT(mEnt);
+  nsresult rv;
+  nsTArray<nsCString> socketTypes;
+  const nsHttpConnectionInfo* ci = mEnt->mConnInfo;
+  if (mIsHttp3) {
+    socketTypes.AppendElement("quic"_ns);
+  } else {
+    if (ci->FirstHopSSL()) {
+      socketTypes.AppendElement("ssl"_ns);
+    } else {
+      const nsCString& defaultType = gHttpHandler->DefaultSocketType();
+      if (!defaultType.IsVoid()) {
+        socketTypes.AppendElement(defaultType);
+      }
+    }
+  }
+
+  nsCOMPtr<nsISocketTransport> socketTransport;
+  nsCOMPtr<nsISocketTransportService> sts;
+
+  sts = services::GetSocketTransportService();
+  if (!sts) {
+    return NS_ERROR_NOT_AVAILABLE;
+  }
+
+  LOG(
+      ("nsHalfOpenSocket::SetupStreams [this=%p ent=%s] "
+       "setup routed transport to origin %s:%d via %s:%d\n",
+       this, ci->HashKey().get(), ci->Origin(), ci->OriginPort(),
+       ci->RoutedHost(), ci->RoutedPort()));
+
+  nsCOMPtr<nsIRoutedSocketTransportService> routedSTS(do_QueryInterface(sts));
+  if (routedSTS) {
+    rv = routedSTS->CreateRoutedTransport(
+        socketTypes, ci->GetOrigin(), ci->OriginPort(), ci->GetRoutedHost(),
+        ci->RoutedPort(), ci->ProxyInfo(), getter_AddRefs(socketTransport));
+  } else {
+    if (!ci->GetRoutedHost().IsEmpty()) {
+      // There is a route requested, but the legacy nsISocketTransportService
+      // can't handle it.
+      // Origin should be reachable on origin host name, so this should
+      // not be a problem - but log it.
+      LOG(
+          ("nsHalfOpenSocket this=%p using legacy nsISocketTransportService "
+           "means explicit route %s:%d will be ignored.\n",
+           this, ci->RoutedHost(), ci->RoutedPort()));
+    }
+
+    rv = sts->CreateTransport(socketTypes, ci->GetOrigin(), ci->OriginPort(),
+                              ci->ProxyInfo(), getter_AddRefs(socketTransport));
+  }
+  NS_ENSURE_SUCCESS(rv, rv);
+
+  uint32_t tmpFlags = 0;
+  if (mCaps & NS_HTTP_REFRESH_DNS) tmpFlags = nsISocketTransport::BYPASS_CACHE;
+
+  tmpFlags |= nsISocketTransport::GetFlagsFromTRRMode(
+      NS_HTTP_TRR_MODE_FROM_FLAGS(mCaps));
+
+  if (mCaps & NS_HTTP_LOAD_ANONYMOUS)
+    tmpFlags |= nsISocketTransport::ANONYMOUS_CONNECT;
+
+  if (ci->GetPrivate() || ci->GetIsolated()) {
+    tmpFlags |= nsISocketTransport::NO_PERMANENT_STORAGE;
+  }
+
+  if (ci->GetLessThanTls13()) {
+    tmpFlags |= nsISocketTransport::DONT_TRY_ESNI_OR_ECH;
+  }
+
+  if (((mCaps & NS_HTTP_BE_CONSERVATIVE) || ci->GetBeConservative()) &&
+      gHttpHandler->ConnMgr()->BeConservativeIfProxied(ci->ProxyInfo())) {
+    LOG(("Setting Socket to BE_CONSERVATIVE"));
+    tmpFlags |= nsISocketTransport::BE_CONSERVATIVE;
+  }
+
+  if (ci->HasIPHintAddress()) {
+    nsCOMPtr<nsIDNSService> dns =
+        do_GetService("@mozilla.org/network/dns-service;1", &rv);
+    if (NS_FAILED(rv)) {
+      return rv;
+    }
+
+    // The spec says: "If A and AAAA records for TargetName are locally
+    // available, the client SHOULD ignore these hints.", so we check if the DNS
+    // record is in cache before setting USE_IP_HINT_ADDRESS.
+    nsCOMPtr<nsIDNSRecord> record;
+    rv = dns->ResolveNative(ci->GetRoutedHost(), nsIDNSService::RESOLVE_OFFLINE,
+                            mEnt->mConnInfo->GetOriginAttributes(),
+                            getter_AddRefs(record));
+    if (NS_FAILED(rv) || !record) {
+      LOG(("Setting Socket to use IP hint address"));
+      tmpFlags |= nsISocketTransport::USE_IP_HINT_ADDRESS;
+    }
+  }
+
+  if (mCaps & NS_HTTP_DISABLE_IPV4) {
+    tmpFlags |= nsISocketTransport::DISABLE_IPV4;
+  } else if (mCaps & NS_HTTP_DISABLE_IPV6) {
+    tmpFlags |= nsISocketTransport::DISABLE_IPV6;
+  } else if (mEnt->PreferenceKnown()) {
+    if (mEnt->mPreferIPv6) {
+      tmpFlags |= nsISocketTransport::DISABLE_IPV4;
+    } else if (mEnt->mPreferIPv4) {
+      tmpFlags |= nsISocketTransport::DISABLE_IPV6;
+    }
+
+    // In case the host is no longer accessible via the preferred IP family,
+    // try the opposite one and potentially restate the preference.
+    tmpFlags |= nsISocketTransport::RETRY_WITH_DIFFERENT_IP_FAMILY;
+
+    // From the same reason, let the backup socket fail faster to try the other
+    // family.
+    uint16_t fallbackTimeout =
+        isBackup ? gHttpHandler->GetFallbackSynTimeout() : 0;
+    if (fallbackTimeout) {
+      socketTransport->SetTimeout(nsISocketTransport::TIMEOUT_CONNECT,
+                                  fallbackTimeout);
+    }
+  } else if (isBackup && gHttpHandler->FastFallbackToIPv4()) {
+    // For backup connections, we disable IPv6. That's because some users have
+    // broken IPv6 connectivity (leading to very long timeouts), and disabling
+    // IPv6 on the backup connection gives them a much better user experience
+    // with dual-stack hosts, though they still pay the 250ms delay for each new
+    // connection. This strategy is also known as "happy eyeballs".
+    tmpFlags |= nsISocketTransport::DISABLE_IPV6;
+  }
+
+  if (!Allow1918()) {
+    tmpFlags |= nsISocketTransport::DISABLE_RFC1918;
+  }
+
+  if ((mFastOpenStatus != TFO_HTTP) && !isBackup) {
+    if (mEnt->mUseFastOpen) {
+      socketTransport->SetFastOpenCallback(this);
+    } else {
+      mFastOpenStatus = TFO_DISABLED;
+    }
+  }
+
+  MOZ_ASSERT(!(tmpFlags & nsISocketTransport::DISABLE_IPV4) ||
+                 !(tmpFlags & nsISocketTransport::DISABLE_IPV6),
+             "Both types should not be disabled at the same time.");
+  socketTransport->SetConnectionFlags(tmpFlags);
+  socketTransport->SetTlsFlags(ci->GetTlsFlags());
+
+  const OriginAttributes& originAttributes =
+      mEnt->mConnInfo->GetOriginAttributes();
+  if (originAttributes != OriginAttributes()) {
+    socketTransport->SetOriginAttributes(originAttributes);
+  }
+
+  socketTransport->SetQoSBits(gHttpHandler->GetQoSBits());
+
+  rv = socketTransport->SetEventSink(this, nullptr);
+  NS_ENSURE_SUCCESS(rv, rv);
+
+  rv = socketTransport->SetSecurityCallbacks(this);
+  NS_ENSURE_SUCCESS(rv, rv);
+
+  if (gHttpHandler->EchConfigEnabled()) {
+    rv = socketTransport->SetEchConfig(ci->GetEchConfig());
+    NS_ENSURE_SUCCESS(rv, rv);
+  }
+
+  Telemetry::Accumulate(Telemetry::HTTP_CONNECTION_ENTRY_CACHE_HIT_1,
+                        mEnt->mUsedForConnection);
+  mEnt->mUsedForConnection = true;
+
+  nsCOMPtr<nsIOutputStream> sout;
+  rv = socketTransport->OpenOutputStream(nsITransport::OPEN_UNBUFFERED, 0, 0,
+                                         getter_AddRefs(sout));
+  NS_ENSURE_SUCCESS(rv, rv);
+
+  nsCOMPtr<nsIInputStream> sin;
+  rv = socketTransport->OpenInputStream(nsITransport::OPEN_UNBUFFERED, 0, 0,
+                                        getter_AddRefs(sin));
+  NS_ENSURE_SUCCESS(rv, rv);
+
+  socketTransport.forget(transport);
+  CallQueryInterface(sin, instream);
+  CallQueryInterface(sout, outstream);
+
+  rv = (*outstream)->AsyncWait(this, 0, 0, nullptr);
+  if (NS_SUCCEEDED(rv)) gHttpHandler->ConnMgr()->StartedConnect();
+
+  return rv;
+}
+
+nsresult nsHttpConnectionMgr::nsHalfOpenSocket::SetupPrimaryStreams() {
+  MOZ_ASSERT(OnSocketThread(), "not on socket thread");
+
+  nsresult rv;
+
+  mPrimarySynStarted = TimeStamp::Now();
+  rv = SetupStreams(getter_AddRefs(mSocketTransport), getter_AddRefs(mStreamIn),
+                    getter_AddRefs(mStreamOut), false);
+
+  LOG(("nsHalfOpenSocket::SetupPrimaryStream [this=%p ent=%s rv=%" PRIx32 "]",
+       this, mEnt->mConnInfo->Origin(), static_cast<uint32_t>(rv)));
+  if (NS_FAILED(rv)) {
+    if (mStreamOut) mStreamOut->AsyncWait(nullptr, 0, 0, nullptr);
+    if (mSocketTransport) {
+      mSocketTransport->SetFastOpenCallback(nullptr);
+    }
+    mStreamOut = nullptr;
+    mStreamIn = nullptr;
+    mSocketTransport = nullptr;
+  }
+  return rv;
+}
+
+nsresult nsHttpConnectionMgr::nsHalfOpenSocket::SetupBackupStreams() {
+  MOZ_ASSERT(mTransaction);
+
+  mBackupSynStarted = TimeStamp::Now();
+  nsresult rv = SetupStreams(getter_AddRefs(mBackupTransport),
+                             getter_AddRefs(mBackupStreamIn),
+                             getter_AddRefs(mBackupStreamOut), true);
+
+  LOG(("nsHalfOpenSocket::SetupBackupStream [this=%p ent=%s rv=%" PRIx32 "]",
+       this, mEnt->mConnInfo->Origin(), static_cast<uint32_t>(rv)));
+  if (NS_FAILED(rv)) {
+    if (mBackupStreamOut) mBackupStreamOut->AsyncWait(nullptr, 0, 0, nullptr);
+    mBackupStreamOut = nullptr;
+    mBackupStreamIn = nullptr;
+    mBackupTransport = nullptr;
+  }
+  return rv;
+}
+
+void nsHttpConnectionMgr::nsHalfOpenSocket::SetupBackupTimer() {
+  MOZ_ASSERT(mEnt);
+  uint16_t timeout = gHttpHandler->GetIdleSynTimeout();
+  MOZ_ASSERT(!mSynTimer, "timer already initd");
+  if (!timeout && mFastOpenInProgress) {
+    timeout = 250;
+  }
+  // When using Fast Open the correct transport will be setup for sure (it is
+  // guaranteed), but it can be that it will happened a bit later.
+  if (mFastOpenInProgress || (timeout && !mSpeculative)) {
+    // Setup the timer that will establish a backup socket
+    // if we do not get a writable event on the main one.
+    // We do this because a lost SYN takes a very long time
+    // to repair at the TCP level.
+    //
+    // Failure to setup the timer is something we can live with,
+    // so don't return an error in that case.
+    NS_NewTimerWithCallback(getter_AddRefs(mSynTimer), this, timeout,
+                            nsITimer::TYPE_ONE_SHOT);
+    LOG(("nsHalfOpenSocket::SetupBackupTimer() [this=%p]", this));
+  } else if (timeout) {
+    LOG(("nsHalfOpenSocket::SetupBackupTimer() [this=%p], did not arm\n",
+         this));
+  }
+}
+
+void nsHttpConnectionMgr::nsHalfOpenSocket::CancelBackupTimer() {
+  // If the syntimer is still armed, we can cancel it because no backup
+  // socket should be formed at this point
+  if (!mSynTimer) return;
+
+  LOG(("nsHalfOpenSocket::CancelBackupTimer()"));
+  mSynTimer->Cancel();
+
+  // Keeping the reference to the timer to remember we have already
+  // performed the backup connection.
+}
+
+void nsHttpConnectionMgr::nsHalfOpenSocket::Abandon() {
+  LOG(("nsHalfOpenSocket::Abandon [this=%p ent=%s] %p %p %p %p", this,
+       mEnt->mConnInfo->Origin(), mSocketTransport.get(),
+       mBackupTransport.get(), mStreamOut.get(), mBackupStreamOut.get()));
+
+  MOZ_ASSERT(OnSocketThread(), "not on socket thread");
+
+  RefPtr<nsHalfOpenSocket> deleteProtector(this);
+
+  // Tell socket (and backup socket) to forget the half open socket.
+  if (mSocketTransport) {
+    mSocketTransport->SetEventSink(nullptr, nullptr);
+    mSocketTransport->SetSecurityCallbacks(nullptr);
+    mSocketTransport->SetFastOpenCallback(nullptr);
+    mSocketTransport = nullptr;
+  }
+  if (mBackupTransport) {
+    mBackupTransport->SetEventSink(nullptr, nullptr);
+    mBackupTransport->SetSecurityCallbacks(nullptr);
+    mBackupTransport = nullptr;
+  }
+
+  // Tell output stream (and backup) to forget the half open socket.
+  if (mStreamOut) {
+    if (!mFastOpenInProgress) {
+      // If mFastOpenInProgress is true HalfOpen are not in mHalfOpen
+      // list and are not counted so we do not need to decrease counter.
+      gHttpHandler->ConnMgr()->RecvdConnect();
+    }
+    mStreamOut->AsyncWait(nullptr, 0, 0, nullptr);
+    mStreamOut = nullptr;
+  }
+  if (mBackupStreamOut) {
+    gHttpHandler->ConnMgr()->RecvdConnect();
+    mBackupStreamOut->AsyncWait(nullptr, 0, 0, nullptr);
+    mBackupStreamOut = nullptr;
+  }
+
+  // Lose references to input stream (and backup).
+  if (mStreamIn) {
+    mStreamIn->AsyncWait(nullptr, 0, 0, nullptr);
+    mStreamIn = nullptr;
+  }
+  if (mBackupStreamIn) {
+    mBackupStreamIn->AsyncWait(nullptr, 0, 0, nullptr);
+    mBackupStreamIn = nullptr;
+  }
+
+  // Stop the timer - we don't want any new backups.
+  CancelBackupTimer();
+
+  // Remove the half open from the connection entry.
+  if (mEnt) {
+    mEnt->mDoNotDestroy = false;
+    mEnt->RemoveHalfOpen(this);
+  }
+  mEnt = nullptr;
+}
+
+double nsHttpConnectionMgr::nsHalfOpenSocket::Duration(TimeStamp epoch) {
+  if (mPrimarySynStarted.IsNull()) return 0;
+
+  return (epoch - mPrimarySynStarted).ToMilliseconds();
+}
+
+NS_IMETHODIMP  // method for nsITimerCallback
+nsHttpConnectionMgr::nsHalfOpenSocket::Notify(nsITimer* timer) {
+  MOZ_ASSERT(OnSocketThread(), "not on socket thread");
+  MOZ_ASSERT(timer == mSynTimer, "wrong timer");
+
+  MOZ_ASSERT(!mBackupTransport);
+  MOZ_ASSERT(mSynTimer);
+  MOZ_ASSERT(mEnt);
+
+  DebugOnly<nsresult> rv = SetupBackupStreams();
+  MOZ_ASSERT(NS_SUCCEEDED(rv));
+
+  // Keeping the reference to the timer to remember we have already
+  // performed the backup connection.
+
+  return NS_OK;
+}
+
+NS_IMETHODIMP  // method for nsINamed
+nsHttpConnectionMgr::nsHalfOpenSocket::GetName(nsACString& aName) {
+  aName.AssignLiteral("nsHttpConnectionMgr::nsHalfOpenSocket");
+  return NS_OK;
+}
+
+already_AddRefed<nsHttpConnectionMgr::PendingTransactionInfo>
+nsHttpConnectionMgr::nsHalfOpenSocket::FindTransactionHelper(
+    bool removeWhenFound) {
+  nsTArray<RefPtr<PendingTransactionInfo>>* pendingQ =
+      gHttpHandler->ConnMgr()->GetTransactionPendingQHelper(mEnt, mTransaction);
+
+  int32_t index =
+      pendingQ ? pendingQ->IndexOf(mTransaction, 0, PendingComparator()) : -1;
+
+  RefPtr<PendingTransactionInfo> info;
+  if (index != -1) {
+    info = (*pendingQ)[index];
+    if (removeWhenFound) {
+      pendingQ->RemoveElementAt(index);
+    }
+  }
+  return info.forget();
+}
+
+// method for nsIAsyncOutputStreamCallback
+NS_IMETHODIMP
+nsHttpConnectionMgr::nsHalfOpenSocket::OnOutputStreamReady(
+    nsIAsyncOutputStream* out) {
+  MOZ_ASSERT(OnSocketThread(), "not on socket thread");
+  MOZ_ASSERT(mStreamOut || mBackupStreamOut);
+  MOZ_ASSERT(out == mStreamOut || out == mBackupStreamOut, "stream mismatch");
+  MOZ_ASSERT(mEnt);
+
+  LOG(("nsHalfOpenSocket::OnOutputStreamReady [this=%p ent=%s %s]\n", this,
+       mEnt->mConnInfo->Origin(), out == mStreamOut ? "primary" : "backup"));
+
+  mEnt->mDoNotDestroy = true;
+  gHttpHandler->ConnMgr()->RecvdConnect();
+
+  CancelBackupTimer();
+
+  if (mFastOpenInProgress) {
+    LOG(
+        ("nsHalfOpenSocket::OnOutputStreamReady backup stream is ready, "
+         "close the fast open socket %p [this=%p ent=%s]\n",
+         mSocketTransport.get(), this, mEnt->mConnInfo->Origin()));
+    // If fast open is used, right after a socket for the primary stream is
+    // created a HttpConnectionBase is created for that socket. The connection
+    // listens for  OnOutputStreamReady not HalfOpenSocket. So this stream
+    // cannot be mStreamOut.
+    MOZ_ASSERT((out == mBackupStreamOut) && mConnectionNegotiatingFastOpen);
+    // Here the backup, non-TFO connection has connected successfully,
+    // before the TFO connection.
+    //
+    // The primary, TFO connection will be cancelled and the transaction
+    // will be rewind. CloseConnectionFastOpenTakesTooLongOrError will
+    // return the rewind transaction. The transaction will be put back to
+    // the pending queue and as well connected to this halfOpenSocket.
+    // SetupConn should set up a new HttpConnectionBase with the backup
+    // socketTransport and the rewind transaction.
+    mSocketTransport->SetFastOpenCallback(nullptr);
+    mConnectionNegotiatingFastOpen->SetFastOpen(false);
+    mEnt->mHalfOpenFastOpenBackups.RemoveElement(this);
+    RefPtr<nsAHttpTransaction> trans =
+        mConnectionNegotiatingFastOpen
+            ->CloseConnectionFastOpenTakesTooLongOrError(true);
+    mSocketTransport = nullptr;
+    mStreamOut = nullptr;
+    mStreamIn = nullptr;
+
+    if (trans && trans->QueryHttpTransaction()) {
+      RefPtr<PendingTransactionInfo> pendingTransInfo =
+          new PendingTransactionInfo(trans->QueryHttpTransaction());
+      pendingTransInfo->mHalfOpen =
+          do_GetWeakReference(static_cast<nsISupportsWeakReference*>(this));
+      if (trans->Caps() & NS_HTTP_URGENT_START) {
+        gHttpHandler->ConnMgr()->InsertTransactionSorted(
+            mEnt->mUrgentStartQ, pendingTransInfo, true);
+      } else {
+        mEnt->InsertTransaction(pendingTransInfo, true);
+      }
+    }
+    if (mEnt->mUseFastOpen) {
+      gHttpHandler->IncrementFastOpenConsecutiveFailureCounter();
+      mEnt->mUseFastOpen = false;
+    }
+
+    mFastOpenInProgress = false;
+    mConnectionNegotiatingFastOpen = nullptr;
+    if (mFastOpenStatus == TFO_NOT_TRIED) {
+      mFastOpenStatus = TFO_FAILED_BACKUP_CONNECTION_TFO_NOT_TRIED;
+    } else if (mFastOpenStatus == TFO_TRIED) {
+      mFastOpenStatus = TFO_FAILED_BACKUP_CONNECTION_TFO_TRIED;
+    } else if (mFastOpenStatus == TFO_DATA_SENT) {
+      mFastOpenStatus = TFO_FAILED_BACKUP_CONNECTION_TFO_DATA_SENT;
+    } else {
+      // This is TFO_DATA_COOKIE_NOT_ACCEPTED (I think this cannot
+      // happened, because the primary connection will be already
+      // connected or in recovery and mFastOpenInProgress==false).
+      mFastOpenStatus =
+          TFO_FAILED_BACKUP_CONNECTION_TFO_DATA_COOKIE_NOT_ACCEPTED;
+    }
+  }
+
+  if (((mFastOpenStatus == TFO_DISABLED) || (mFastOpenStatus == TFO_HTTP)) &&
+      !mBackupConnStatsSet) {
+    // Collect telemetry for backup connection being faster than primary
+    // connection. We want to collect this telemetry only for cases where
+    // TFO is not used.
+    mBackupConnStatsSet = true;
+    Telemetry::Accumulate(Telemetry::NETWORK_HTTP_BACKUP_CONN_WON_1,
+                          (out == mBackupStreamOut));
+  }
+
+  if (mFastOpenStatus == TFO_UNKNOWN) {
+    MOZ_ASSERT(out == mStreamOut);
+    if (mPrimaryStreamStatus == NS_NET_STATUS_RESOLVING_HOST) {
+      mFastOpenStatus = TFO_UNKNOWN_RESOLVING;
+    } else if (mPrimaryStreamStatus == NS_NET_STATUS_RESOLVED_HOST) {
+      mFastOpenStatus = TFO_UNKNOWN_RESOLVED;
+    } else if (mPrimaryStreamStatus == NS_NET_STATUS_CONNECTING_TO) {
+      mFastOpenStatus = TFO_UNKNOWN_CONNECTING;
+    } else if (mPrimaryStreamStatus == NS_NET_STATUS_CONNECTED_TO) {
+      mFastOpenStatus = TFO_UNKNOWN_CONNECTED;
+    }
+  }
+  nsresult rv = SetupConn(out, false);
+  if (mEnt) {
+    mEnt->mDoNotDestroy = false;
+  }
+  return rv;
+}
+
+bool nsHttpConnectionMgr::nsHalfOpenSocket::FastOpenEnabled() {
+  LOG(("nsHalfOpenSocket::FastOpenEnabled [this=%p]\n", this));
+
+  MOZ_ASSERT(mEnt);
+
+  if (!mEnt) {
+    return false;
+  }
+
+  MOZ_ASSERT(mEnt->mConnInfo->FirstHopSSL());
+
+  // If mEnt is present this HalfOpen must be in the mHalfOpens,
+  // but we want to be sure!!!
+  if (!mEnt->mHalfOpens.Contains(this)) {
+    return false;
+  }
+
+  if (!gHttpHandler->UseFastOpen()) {
+    // fast open was turned off.
+    LOG(("nsHalfOpenSocket::FastEnabled - fast open was turned off.\n"));
+    mEnt->mUseFastOpen = false;
+    mFastOpenStatus = TFO_DISABLED;
+    return false;
+  }
+  // We can use FastOpen if we have a transaction or if it is ssl
+  // connection. For ssl we will use a null transaction to drive the SSL
+  // handshake to completion if there is not a pending transaction. Afterwards
+  // the connection will be 100% ready for the next transaction to use it.
+  // Make an exception for SSL tunneled HTTP proxy as the NullHttpTransaction
+  // does not know how to drive Connect.
+  if (mEnt->mConnInfo->UsingConnect()) {
+    LOG(("nsHalfOpenSocket::FastOpenEnabled - It is using Connect."));
+    mFastOpenStatus = TFO_DISABLED_CONNECT;
+    return false;
+  }
+  return true;
+}
+
+nsresult nsHttpConnectionMgr::nsHalfOpenSocket::StartFastOpen() {
+  MOZ_ASSERT(mStreamOut);
+  MOZ_ASSERT(!mBackupTransport);
+  MOZ_ASSERT(mEnt);
+  MOZ_ASSERT(mFastOpenStatus == TFO_UNKNOWN);
+
+  LOG(("nsHalfOpenSocket::StartFastOpen [this=%p]\n", this));
+
+  RefPtr<nsHalfOpenSocket> deleteProtector(this);
+
+  mFastOpenInProgress = true;
+  mEnt->mDoNotDestroy = true;
+  // Remove this HalfOpen from mEnt->mHalfOpens.
+  // The new connection will take care of closing this HalfOpen from now on!
+  if (!mEnt->mHalfOpens.RemoveElement(this)) {
+    MOZ_ASSERT(false, "HalfOpen is not in mHalfOpens!");
+    mSocketTransport->SetFastOpenCallback(nullptr);
+    CancelBackupTimer();
+    mFastOpenInProgress = false;
+    Abandon();
+    mFastOpenStatus = TFO_INIT_FAILED;
+    return NS_ERROR_ABORT;
+  }
+
+  MOZ_ASSERT(gHttpHandler->ConnMgr()->mNumHalfOpenConns);
+  if (gHttpHandler->ConnMgr()->mNumHalfOpenConns) {  // just in case
+    gHttpHandler->ConnMgr()->mNumHalfOpenConns--;
+  }
+
+  // Count this socketTransport as connected.
+  gHttpHandler->ConnMgr()->RecvdConnect();
+
+  // Remove HalfOpen from callbacks, the new connection will take them.
+  mSocketTransport->SetEventSink(nullptr, nullptr);
+  mSocketTransport->SetSecurityCallbacks(nullptr);
+  mStreamOut->AsyncWait(nullptr, 0, 0, nullptr);
+
+  nsresult rv = SetupConn(mStreamOut, true);
+  if (!mConnectionNegotiatingFastOpen) {
+    LOG(
+        ("nsHalfOpenSocket::StartFastOpen SetupConn failed "
+         "[this=%p rv=%x]\n",
+         this, static_cast<uint32_t>(rv)));
+    if (NS_SUCCEEDED(rv)) {
+      rv = NS_ERROR_ABORT;
+    }
+    // If SetupConn failed this will CloseTransaction and socketTransport
+    // with an error, therefore we can close this HalfOpen. socketTransport
+    // will remove reference to this HalfOpen as well.
+    mSocketTransport->SetFastOpenCallback(nullptr);
+    CancelBackupTimer();
+    mFastOpenInProgress = false;
+
+    // The connection is responsible to take care of the halfOpen so we
+    // need to clean it up.
+    Abandon();
+    mFastOpenStatus = TFO_INIT_FAILED;
+  } else {
+    LOG(("nsHalfOpenSocket::StartFastOpen [this=%p conn=%p]\n", this,
+         mConnectionNegotiatingFastOpen.get()));
+
+    mEnt->mHalfOpenFastOpenBackups.AppendElement(this);
+    // SetupBackupTimer should setup timer which will hold a ref to this
+    // halfOpen. It will failed only if it cannot create timer. Anyway just
+    // to be sure I will add this deleteProtector!!!
+    if (!mSynTimer) {
+      // For Fast Open we will setup backup timer also for
+      // NullTransaction.
+      // So maybe it is not set and we need to set it here.
+      SetupBackupTimer();
+    }
+  }
+  if (mEnt) {
+    mEnt->mDoNotDestroy = false;
+  }
+  return rv;
+}
+
+void nsHttpConnectionMgr::nsHalfOpenSocket::SetFastOpenConnected(
+    nsresult aError, bool aWillRetry) {
+  MOZ_ASSERT(mFastOpenInProgress);
+  MOZ_ASSERT(mEnt);
+
+  LOG(("nsHalfOpenSocket::SetFastOpenConnected [this=%p conn=%p error=%x]\n",
+       this, mConnectionNegotiatingFastOpen.get(),
+       static_cast<uint32_t>(aError)));
+
+  // mConnectionNegotiatingFastOpen is set after a StartFastOpen creates
+  // and activates a HttpConnectionBase successfully (SetupConn calls
+  // DispatchTransaction and DispatchAbstractTransaction which calls
+  // conn->Activate).
+  // HttpConnectionBase::Activate can fail which will close socketTransport
+  // and socketTransport will call this function. The FastOpen clean up
+  // in case HttpConnectionBase::Activate fails will be done in StartFastOpen.
+  // Also OnMsgReclaimConnection can decided that we do not need this
+  // transaction and cancel it as well.
+  // In all other cases mConnectionNegotiatingFastOpen must not be nullptr.
+  if (!mConnectionNegotiatingFastOpen) {
+    return;
+  }
+
+  MOZ_ASSERT((mFastOpenStatus == TFO_NOT_TRIED) ||
+             (mFastOpenStatus == TFO_DATA_SENT) ||
+             (mFastOpenStatus == TFO_TRIED) ||
+             (mFastOpenStatus == TFO_DATA_COOKIE_NOT_ACCEPTED) ||
+             (mFastOpenStatus == TFO_DISABLED));
+
+  RefPtr<nsHalfOpenSocket> deleteProtector(this);
+
+  mEnt->mDoNotDestroy = true;
+
+  // Delete 2 points of entry to FastOpen function so that we do not reenter.
+  mEnt->mHalfOpenFastOpenBackups.RemoveElement(this);
+  mSocketTransport->SetFastOpenCallback(nullptr);
+
+  mConnectionNegotiatingFastOpen->SetFastOpen(false);
+
+  // Check if we want to restart connection!
+  if (aWillRetry && ((aError == NS_ERROR_CONNECTION_REFUSED) ||
+#if defined(_WIN64) && defined(WIN95)
+                     // On Windows PR_ContinueConnect can return
+                     // NS_ERROR_FAILURE. This will be fixed in bug 1386719 and
+                     // this is just a temporary work around.
+                     (aError == NS_ERROR_FAILURE) ||
+#endif
+                     (aError == NS_ERROR_PROXY_CONNECTION_REFUSED) ||
+                     (aError == NS_ERROR_NET_TIMEOUT))) {
+    if (mEnt->mUseFastOpen) {
+      gHttpHandler->IncrementFastOpenConsecutiveFailureCounter();
+      mEnt->mUseFastOpen = false;
+    }
+    // This is called from nsSocketTransport::RecoverFromError. The
+    // socket will try connect and we need to rewind nsHttpTransaction.
+
+    RefPtr<nsAHttpTransaction> trans =
+        mConnectionNegotiatingFastOpen
+            ->CloseConnectionFastOpenTakesTooLongOrError(false);
+    if (trans && trans->QueryHttpTransaction()) {
+      RefPtr<PendingTransactionInfo> pendingTransInfo =
+          new PendingTransactionInfo(trans->QueryHttpTransaction());
+      pendingTransInfo->mHalfOpen =
+          do_GetWeakReference(static_cast<nsISupportsWeakReference*>(this));
+      if (trans->Caps() & NS_HTTP_URGENT_START) {
+        gHttpHandler->ConnMgr()->InsertTransactionSorted(
+            mEnt->mUrgentStartQ, pendingTransInfo, true);
+      } else {
+        mEnt->InsertTransaction(pendingTransInfo, true);
+      }
+    }
+    // We are doing a restart without fast open, so the easiest way is to
+    // return mSocketTransport to the halfOpenSock and destroy connection.
+    // This makes http2 implemenntation easier.
+    // mConnectionNegotiatingFastOpen is going away and halfOpen is taking
+    // this mSocketTransport so add halfOpen to mEnt and update
+    // mNumActiveConns.
+    mEnt->mHalfOpens.AppendElement(this);
+    gHttpHandler->ConnMgr()->mNumHalfOpenConns++;
+    gHttpHandler->ConnMgr()->StartedConnect();
+
+    // Restore callbacks.
+    mStreamOut->AsyncWait(this, 0, 0, nullptr);
+    mSocketTransport->SetEventSink(this, nullptr);
+    mSocketTransport->SetSecurityCallbacks(this);
+    mStreamIn->AsyncWait(nullptr, 0, 0, nullptr);
+
+    if ((aError == NS_ERROR_CONNECTION_REFUSED) ||
+        (aError == NS_ERROR_PROXY_CONNECTION_REFUSED)) {
+      mFastOpenStatus = TFO_FAILED_CONNECTION_REFUSED;
+    } else if (aError == NS_ERROR_NET_TIMEOUT) {
+      mFastOpenStatus = TFO_FAILED_NET_TIMEOUT;
+    } else {
+      mFastOpenStatus = TFO_FAILED_UNKNOW_ERROR;
+    }
+
+  } else {
+    // On success or other error we proceed with connection, we just need
+    // to close backup timer and halfOpenSock.
+    CancelBackupTimer();
+    if (NS_SUCCEEDED(aError)) {
+      NetAddr peeraddr;
+      if (NS_SUCCEEDED(mSocketTransport->GetPeerAddr(&peeraddr))) {
+        mEnt->RecordIPFamilyPreference(peeraddr.raw.family);
+      }
+      gHttpHandler->ResetFastOpenConsecutiveFailureCounter();
+    }
+    mSocketTransport = nullptr;
+    mStreamOut = nullptr;
+    mStreamIn = nullptr;
+
+    // If backup transport has already started put this HalfOpen back to
+    // mEnt list.
+    if (mBackupTransport) {
+      mFastOpenStatus = TFO_BACKUP_CONN;
+      mEnt->mHalfOpens.AppendElement(this);
+      gHttpHandler->ConnMgr()->mNumHalfOpenConns++;
+    }
+  }
+
+  mFastOpenInProgress = false;
+  mConnectionNegotiatingFastOpen = nullptr;
+  if (mEnt) {
+    mEnt->mDoNotDestroy = false;
+  } else {
+    MOZ_ASSERT(!mBackupTransport);
+    MOZ_ASSERT(!mBackupStreamOut);
+  }
+}
+
+void nsHttpConnectionMgr::nsHalfOpenSocket::SetFastOpenStatus(
+    uint8_t tfoStatus) {
+  MOZ_ASSERT(OnSocketThread(), "not on socket thread");
+  MOZ_ASSERT(mFastOpenInProgress);
+
+  mFastOpenStatus = tfoStatus;
+  mConnectionNegotiatingFastOpen->SetFastOpenStatus(tfoStatus);
+  if (mConnectionNegotiatingFastOpen->Transaction()) {
+    // The transaction could already be canceled in the meantime, hence
+    // nullified.
+    mConnectionNegotiatingFastOpen->Transaction()->SetFastOpenStatus(tfoStatus);
+  }
+}
+
+void nsHttpConnectionMgr::nsHalfOpenSocket::CancelFastOpenConnection() {
+  MOZ_ASSERT(mFastOpenInProgress);
+
+  LOG(("nsHalfOpenSocket::CancelFastOpenConnection [this=%p conn=%p]\n", this,
+       mConnectionNegotiatingFastOpen.get()));
+
+  RefPtr<nsHalfOpenSocket> deleteProtector(this);
+  mEnt->mHalfOpenFastOpenBackups.RemoveElement(this);
+  mSocketTransport->SetFastOpenCallback(nullptr);
+  mConnectionNegotiatingFastOpen->SetFastOpen(false);
+  RefPtr<nsAHttpTransaction> trans =
+      mConnectionNegotiatingFastOpen
+          ->CloseConnectionFastOpenTakesTooLongOrError(true);
+  mSocketTransport = nullptr;
+  mStreamOut = nullptr;
+  mStreamIn = nullptr;
+
+  if (trans && trans->QueryHttpTransaction()) {
+    RefPtr<PendingTransactionInfo> pendingTransInfo =
+        new PendingTransactionInfo(trans->QueryHttpTransaction());
+
+    if (trans->Caps() & NS_HTTP_URGENT_START) {
+      gHttpHandler->ConnMgr()->InsertTransactionSorted(mEnt->mUrgentStartQ,
+                                                       pendingTransInfo, true);
+    } else {
+      mEnt->InsertTransaction(pendingTransInfo, true);
+    }
+  }
+
+  mFastOpenInProgress = false;
+  mConnectionNegotiatingFastOpen = nullptr;
+  Abandon();
+
+  MOZ_ASSERT(!mBackupTransport);
+  MOZ_ASSERT(!mBackupStreamOut);
+}
+
+void nsHttpConnectionMgr::nsHalfOpenSocket::FastOpenNotSupported() {
+  MOZ_ASSERT(mFastOpenInProgress);
+  gHttpHandler->SetFastOpenNotSupported();
+}
+
+nsresult nsHttpConnectionMgr::nsHalfOpenSocket::SetupConn(
+    nsIAsyncOutputStream* out, bool aFastOpen) {
+  MOZ_ASSERT(!aFastOpen || (out == mStreamOut));
+  // We cannot ask for a connection for TFO and Http3 ata the same time.
+  MOZ_ASSERT(!(mIsHttp3 && aFastOpen));
+  // assign the new socket to the http connection
+  RefPtr<HttpConnectionBase> conn;
+  if (!mIsHttp3) {
+    conn = new nsHttpConnection();
+  } else {
+    conn = new HttpConnectionUDP();
+  }
+
+  LOG(
+      ("nsHalfOpenSocket::SetupConn "
+       "Created new nshttpconnection %p %s\n",
+       conn.get(), mIsHttp3 ? "using http3" : ""));
+
+  NullHttpTransaction* nullTrans = mTransaction->QueryNullTransaction();
+  if (nullTrans) {
+    conn->BootstrapTimings(nullTrans->Timings());
+  }
+
+  // Some capabilities are needed before a transaciton actually gets
+  // scheduled (e.g. how to negotiate false start)
+  conn->SetTransactionCaps(mTransaction->Caps());
+
+  NetAddr peeraddr;
+  nsCOMPtr<nsIInterfaceRequestor> callbacks;
+  mTransaction->GetSecurityCallbacks(getter_AddRefs(callbacks));
+  nsresult rv;
+  if (out == mStreamOut) {
+    TimeDuration rtt = TimeStamp::Now() - mPrimarySynStarted;
+    rv = conn->Init(
+        mEnt->mConnInfo, gHttpHandler->ConnMgr()->mMaxRequestDelay,
+        mSocketTransport, mStreamIn, mStreamOut,
+        mPrimaryConnectedOK || aFastOpen, callbacks,
+        PR_MillisecondsToInterval(static_cast<uint32_t>(rtt.ToMilliseconds())));
+
+    bool resetPreference = false;
+    mSocketTransport->GetResetIPFamilyPreference(&resetPreference);
+    if (resetPreference) {
+      mEnt->ResetIPFamilyPreference();
+    }
+
+    if (!aFastOpen && NS_SUCCEEDED(mSocketTransport->GetPeerAddr(&peeraddr))) {
+      mEnt->RecordIPFamilyPreference(peeraddr.raw.family);
+    }
+
+    // The nsHttpConnection object now owns these streams and sockets
+    if (!aFastOpen) {
+      mStreamOut = nullptr;
+      mStreamIn = nullptr;
+      mSocketTransport = nullptr;
+    } else {
+      RefPtr<nsHttpConnection> connTCP = do_QueryObject(conn);
+      MOZ_ASSERT(connTCP);
+      if (connTCP) {
+        connTCP->SetFastOpen(true);
+      }
+    }
+  } else if (out == mBackupStreamOut) {
+    TimeDuration rtt = TimeStamp::Now() - mBackupSynStarted;
+    rv = conn->Init(
+        mEnt->mConnInfo, gHttpHandler->ConnMgr()->mMaxRequestDelay,
+        mBackupTransport, mBackupStreamIn, mBackupStreamOut, mBackupConnectedOK,
+        callbacks,
+        PR_MillisecondsToInterval(static_cast<uint32_t>(rtt.ToMilliseconds())));
+
+    bool resetPreference = false;
+    mBackupTransport->GetResetIPFamilyPreference(&resetPreference);
+    if (resetPreference) {
+      mEnt->ResetIPFamilyPreference();
+    }
+
+    if (NS_SUCCEEDED(mBackupTransport->GetPeerAddr(&peeraddr))) {
+      mEnt->RecordIPFamilyPreference(peeraddr.raw.family);
+    }
+
+    // The nsHttpConnection object now owns these streams and sockets
+    mBackupStreamOut = nullptr;
+    mBackupStreamIn = nullptr;
+    mBackupTransport = nullptr;
+  } else {
+    MOZ_ASSERT(false, "unexpected stream");
+    rv = NS_ERROR_UNEXPECTED;
+  }
+
+  if (NS_FAILED(rv)) {
+    LOG(
+        ("nsHalfOpenSocket::SetupConn "
+         "conn->init (%p) failed %" PRIx32 "\n",
+         conn.get(), static_cast<uint32_t>(rv)));
+
+    RefPtr<nsHttpConnection> connTCP = do_QueryObject(conn);
+    if (connTCP) {
+      // Set TFO status.
+      if ((mFastOpenStatus == TFO_HTTP) || (mFastOpenStatus == TFO_DISABLED) ||
+          (mFastOpenStatus == TFO_DISABLED_CONNECT)) {
+        connTCP->SetFastOpenStatus(mFastOpenStatus);
+      } else {
+        connTCP->SetFastOpenStatus(TFO_INIT_FAILED);
+      }
+    }
+
+    if (nullTrans) {
+      nullTrans->Close(rv);
+    } else if (nsHttpTransaction* trans =
+                   mTransaction->QueryHttpTransaction()) {
+      if (mIsHttp3) {
+        trans->DisableHttp3();
+        gHttpHandler->ExcludeHttp3(mEnt->mConnInfo);
+      }
+      Unused << gHttpHandler->ConnMgr()->CancelTransaction(trans, rv);
+    }
+
+    return rv;
+  }
+
+  // This half-open socket has created a connection.  This flag excludes it
+  // from counter of actual connections used for checking limits.
+  if (!aFastOpen) {
+    mHasConnected = true;
+  }
+
+  // if this is still in the pending list, remove it and dispatch it
+  RefPtr<PendingTransactionInfo> pendingTransInfo = FindTransactionHelper(true);
+  if (pendingTransInfo) {
+    MOZ_ASSERT(!mSpeculative, "Speculative Half Open found mTransaction");
+
+    gHttpHandler->ConnMgr()->AddActiveConn(conn, mEnt);
+    if (mIsHttp3) {
+      // Each connection must have a ConnectionHandle wrapper.
+      // In case of Http < 2 the a ConnectionHandle is created for each
+      // transaction in DispatchAbstractTransaction.
+      // In case of Http2/ and Http3, ConnectionHandle is created only once.
+      // Http2 connection always starts as http1 connection and the first
+      // transaction use DispatchAbstractTransaction to be dispatched and
+      // a ConnectionHandle is created. All consecutive transactions for
+      // Http2 use a short-cut in DispatchTransaction and call
+      // HttpConnectionBase::Activate (DispatchAbstractTransaction is never
+      // called).
+      // In case of Http3 the short-cut HttpConnectionBase::Activate is always
+      // used also for the first transaction, therefore we need to create
+      // ConnectionHandle here.
+      RefPtr<ConnectionHandle> handle = new ConnectionHandle(conn);
+      pendingTransInfo->mTransaction->SetConnection(handle);
+    }
+    rv = gHttpHandler->ConnMgr()->DispatchTransaction(
+        mEnt, pendingTransInfo->mTransaction, conn);
+  } else {
+    // this transaction was dispatched off the pending q before all the
+    // sockets established themselves.
+
+    // After about 1 second allow for the possibility of restarting a
+    // transaction due to server close. Keep at sub 1 second as that is the
+    // minimum granularity we can expect a server to be timing out with.
+    RefPtr<nsHttpConnection> connTCP = do_QueryObject(conn);
+    if (connTCP) {
+      connTCP->SetIsReusedAfter(950);
+    }
+
+    // if we are using ssl and no other transactions are waiting right now,
+    // then form a null transaction to drive the SSL handshake to
+    // completion. Afterwards the connection will be 100% ready for the next
+    // transaction to use it. Make an exception for SSL tunneled HTTP proxy as
+    // the NullHttpTransaction does not know how to drive Connect
+    // Http3 cannot be dispatched using OnMsgReclaimConnection (see below),
+    // therefore we need to use a Nulltransaction.
+    if (!connTCP ||
+        (mEnt->mConnInfo->FirstHopSSL() && !mEnt->mUrgentStartQ.Length() &&
+         !mEnt->PendingQLength() && !mEnt->mConnInfo->UsingConnect())) {
+      LOG(
+          ("nsHalfOpenSocket::SetupConn null transaction will "
+           "be used to finish SSL handshake on conn %p\n",
+           conn.get()));
+      RefPtr<nsAHttpTransaction> trans;
+      if (mTransaction->IsNullTransaction() && !mDispatchedMTransaction) {
+        // null transactions cannot be put in the entry queue, so that
+        // explains why it is not present.
+        mDispatchedMTransaction = true;
+        trans = mTransaction;
+      } else {
+        trans = new NullHttpTransaction(mEnt->mConnInfo, callbacks, mCaps);
+      }
+
+      gHttpHandler->ConnMgr()->AddActiveConn(conn, mEnt);
+      rv = gHttpHandler->ConnMgr()->DispatchAbstractTransaction(mEnt, trans,
+                                                                mCaps, conn, 0);
+    } else {
+      // otherwise just put this in the persistent connection pool
+      LOG(
+          ("nsHalfOpenSocket::SetupConn no transaction match "
+           "returning conn %p to pool\n",
+           conn.get()));
+      gHttpHandler->ConnMgr()->OnMsgReclaimConnection(conn);
+
+      // We expect that there is at least one tranasction in the pending
+      // queue that can take this connection, but it can happened that
+      // all transactions are blocked or they have took other idle
+      // connections. In that case the connection has been added to the
+      // idle queue.
+      // If the connection is in the idle queue but it is using ssl, make
+      // a nulltransaction for it to finish ssl handshake!
+
+      // !!! It can be that mEnt is null after OnMsgReclaimConnection.!!!
+      if (mEnt && mEnt->mConnInfo->FirstHopSSL() &&
+          !mEnt->mConnInfo->UsingConnect()) {
+        int32_t idx = mEnt->mIdleConns.IndexOf(conn);
+        if (idx != -1) {
+          RefPtr<nsHttpConnection> connTCP = do_QueryObject(conn);
+          MOZ_ASSERT(connTCP);
+          if (connTCP) {
+            DebugOnly<nsresult> rvDeb =
+                gHttpHandler->ConnMgr()->RemoveIdleConnection(connTCP);
+            MOZ_ASSERT(NS_SUCCEEDED(rvDeb));
+            connTCP->EndIdleMonitoring();
+          }
+          RefPtr<nsAHttpTransaction> trans;
+          if (mTransaction->IsNullTransaction() && !mDispatchedMTransaction) {
+            mDispatchedMTransaction = true;
+            trans = mTransaction;
+          } else {
+            trans = new NullHttpTransaction(mEnt->mConnInfo, callbacks, mCaps);
+          }
+          gHttpHandler->ConnMgr()->AddActiveConn(conn, mEnt);
+          rv = gHttpHandler->ConnMgr()->DispatchAbstractTransaction(
+              mEnt, trans, mCaps, conn, 0);
+        }
+      }
+    }
+  }
+
+  // If this connection has a transaction get reference to its
+  // ConnectionHandler.
+  RefPtr<nsHttpConnection> connTCP = do_QueryObject(conn);
+  if (connTCP) {
+    if (aFastOpen) {
+      MOZ_ASSERT(mEnt);
+      MOZ_ASSERT(static_cast<int32_t>(mEnt->mIdleConns.IndexOf(connTCP)) == -1);
+      int32_t idx = mEnt->mActiveConns.IndexOf(conn);
+      if (NS_SUCCEEDED(rv) && (idx != -1)) {
+        mConnectionNegotiatingFastOpen = connTCP;
+      } else {
+        connTCP->SetFastOpen(false);
+        connTCP->SetFastOpenStatus(TFO_INIT_FAILED);
+      }
+    } else {
+      connTCP->SetFastOpenStatus(mFastOpenStatus);
+      if ((mFastOpenStatus != TFO_HTTP) && (mFastOpenStatus != TFO_DISABLED) &&
+          (mFastOpenStatus != TFO_DISABLED_CONNECT)) {
+        mFastOpenStatus = TFO_BACKUP_CONN;  // Set this to TFO_BACKUP_CONN
+                                            // so that if a backup
+                                            // connection is established we
+                                            // do not report values twice.
+      }
+    }
+  }
+
+  // If this halfOpenConn was speculative, but at the end the conn got a
+  // non-null transaction than this halfOpen is not speculative anymore!
+  if (conn->Transaction() && !conn->Transaction()->IsNullTransaction()) {
+    Claim();
+  }
+
+  return rv;
 }
 
 // register a connection to receive CanJoinConnection() for particular
@@ -3333,14 +5251,277 @@ void nsHttpConnectionMgr::RegisterOriginCoalescingKey(HttpConnectionBase* conn,
        newKey.get(), conn, ci->HashKey().get()));
 }
 
+// method for nsITransportEventSink
+NS_IMETHODIMP
+nsHttpConnectionMgr::nsHalfOpenSocket::OnTransportStatus(nsITransport* trans,
+                                                         nsresult status,
+                                                         int64_t progress,
+                                                         int64_t progressMax) {
+  MOZ_ASSERT(OnSocketThread(), "not on socket thread");
+
+  MOZ_ASSERT((trans == mSocketTransport) || (trans == mBackupTransport));
+  MOZ_ASSERT(mEnt);
+  if (mTransaction) {
+    if ((trans == mSocketTransport) ||
+        ((trans == mBackupTransport) &&
+         (status == NS_NET_STATUS_CONNECTED_TO) && mSocketTransport)) {
+      // Send this status event only if the transaction is still pending,
+      // i.e. it has not found a free already connected socket.
+      // Sockets in halfOpen state can only get following events:
+      // NS_NET_STATUS_RESOLVING_HOST, NS_NET_STATUS_RESOLVED_HOST,
+      // NS_NET_STATUS_CONNECTING_TO and NS_NET_STATUS_CONNECTED_TO.
+      // mBackupTransport is only started after
+      // NS_NET_STATUS_CONNECTING_TO of mSocketTransport, so ignore all
+      // mBackupTransport events until NS_NET_STATUS_CONNECTED_TO.
+      // mBackupTransport must be connected before mSocketTransport.
+      mTransaction->OnTransportStatus(trans, status, progress);
+    }
+  }
+
+  MOZ_ASSERT(trans == mSocketTransport || trans == mBackupTransport);
+  if (status == NS_NET_STATUS_CONNECTED_TO) {
+    if (trans == mSocketTransport) {
+      mPrimaryConnectedOK = true;
+    } else {
+      mBackupConnectedOK = true;
+    }
+  }
+
+  // The rest of this method only applies to the primary transport
+  if (trans != mSocketTransport) {
+    return NS_OK;
+  }
+
+  mPrimaryStreamStatus = status;
+
+  // if we are doing spdy coalescing and haven't recorded the ip address
+  // for this entry before then make the hash key if our dns lookup
+  // just completed. We can't do coalescing if using a proxy because the
+  // ip addresses are not available to the client.
+
+  if (status == NS_NET_STATUS_CONNECTING_TO && gHttpHandler->IsSpdyEnabled() &&
+      gHttpHandler->CoalesceSpdy() && mEnt && mEnt->mConnInfo &&
+      mEnt->mConnInfo->EndToEndSSL() && mEnt->AllowHttp2() &&
+      !mEnt->mConnInfo->UsingProxy() && mEnt->mCoalescingKeys.IsEmpty()) {
+    nsCOMPtr<nsIDNSAddrRecord> dnsRecord(do_GetInterface(mSocketTransport));
+    nsTArray<NetAddr> addressSet;
+    nsresult rv = NS_ERROR_NOT_AVAILABLE;
+    if (dnsRecord) {
+      rv = dnsRecord->GetAddresses(addressSet);
+    }
+
+    if (NS_SUCCEEDED(rv) && !addressSet.IsEmpty()) {
+      for (uint32_t i = 0; i < addressSet.Length(); ++i) {
+        nsCString* newKey = mEnt->mCoalescingKeys.AppendElement(nsCString());
+        newKey->SetLength(kIPv6CStrBufSize + 26);
+        addressSet[i].ToStringBuffer(newKey->BeginWriting(), kIPv6CStrBufSize);
+        newKey->SetLength(strlen(newKey->BeginReading()));
+        if (mEnt->mConnInfo->GetAnonymous()) {
+          newKey->AppendLiteral("~A:");
+        } else {
+          newKey->AppendLiteral("~.:");
+        }
+        newKey->AppendInt(mEnt->mConnInfo->OriginPort());
+        newKey->AppendLiteral("/[");
+        nsAutoCString suffix;
+        mEnt->mConnInfo->GetOriginAttributes().CreateSuffix(suffix);
+        newKey->Append(suffix);
+        newKey->AppendLiteral("]viaDNS");
+        LOG((
+            "nsHttpConnectionMgr::nsHalfOpenSocket::OnTransportStatus "
+            "STATUS_CONNECTING_TO Established New Coalescing Key # %d for host "
+            "%s [%s]",
+            i, mEnt->mConnInfo->Origin(), newKey->get()));
+      }
+      gHttpHandler->ConnMgr()->ProcessSpdyPendingQ(mEnt);
+    }
+  }
+
+  switch (status) {
+    case NS_NET_STATUS_CONNECTING_TO:
+      // Passed DNS resolution, now trying to connect, start the backup timer
+      // only prevent creating another backup transport.
+      // We also check for mEnt presence to not instantiate the timer after
+      // this half open socket has already been abandoned.  It may happen
+      // when we get this notification right between main-thread calls to
+      // nsHttpConnectionMgr::Shutdown and nsSocketTransportService::Shutdown
+      // where the first abandons all half open socket instances and only
+      // after that the second stops the socket thread.
+      // Http3 has its own syn-retransmission, therefore it does not need a
+      // backup connection.
+      if (mEnt && !mBackupTransport && !mSynTimer && !mIsHttp3) {
+        SetupBackupTimer();
+      }
+      break;
+
+    case NS_NET_STATUS_CONNECTED_TO:
+      // TCP connection's up, now transfer or SSL negotiantion starts,
+      // no need for backup socket
+      CancelBackupTimer();
+      break;
+
+    default:
+      break;
+  }
+
+  return NS_OK;
+}
+
+// method for nsIInterfaceRequestor
+NS_IMETHODIMP
+nsHttpConnectionMgr::nsHalfOpenSocket::GetInterface(const nsIID& iid,
+                                                    void** result) {
+  if (mTransaction) {
+    nsCOMPtr<nsIInterfaceRequestor> callbacks;
+    mTransaction->GetSecurityCallbacks(getter_AddRefs(callbacks));
+    if (callbacks) return callbacks->GetInterface(iid, result);
+  }
+  return NS_ERROR_NO_INTERFACE;
+}
+
+bool nsHttpConnectionMgr::nsHalfOpenSocket::AcceptsTransaction(
+    nsHttpTransaction* trans) {
+  // When marked as urgent start, only accept urgent start marked transactions.
+  // Otherwise, accept any kind of transaction.
+  return !mUrgentStart || (trans->Caps() & nsIClassOfService::UrgentStart);
+}
+
+bool nsHttpConnectionMgr::nsHalfOpenSocket::Claim() {
+  if (mSpeculative) {
+    mSpeculative = false;
+    uint32_t flags;
+    if (mSocketTransport &&
+        NS_SUCCEEDED(mSocketTransport->GetConnectionFlags(&flags))) {
+      flags &= ~nsISocketTransport::DISABLE_RFC1918;
+      mSocketTransport->SetConnectionFlags(flags);
+    }
+
+    Telemetry::AutoCounter<Telemetry::HTTPCONNMGR_USED_SPECULATIVE_CONN>
+        usedSpeculativeConn;
+    ++usedSpeculativeConn;
+
+    if (mIsFromPredictor) {
+      Telemetry::AutoCounter<Telemetry::PREDICTOR_TOTAL_PRECONNECTS_USED>
+          totalPreconnectsUsed;
+      ++totalPreconnectsUsed;
+    }
+
+    // Http3 has its own syn-retransmission, therefore it does not need a
+    // backup connection.
+    if ((mPrimaryStreamStatus == NS_NET_STATUS_CONNECTING_TO) && mEnt &&
+        !mBackupTransport && !mSynTimer && !mIsHttp3) {
+      SetupBackupTimer();
+    }
+  }
+
+  if (mFreeToUse) {
+    mFreeToUse = false;
+    return true;
+  }
+  return false;
+}
+
+void nsHttpConnectionMgr::nsHalfOpenSocket::Unclaim() {
+  MOZ_ASSERT(!mSpeculative && !mFreeToUse);
+  // We will keep the backup-timer running. Most probably this halfOpen will
+  // be used by a transaction from which this transaction took the halfOpen.
+  // (this is happening because of the transaction priority.)
+  mFreeToUse = true;
+}
+
+already_AddRefed<HttpConnectionBase> ConnectionHandle::TakeHttpConnection() {
+  // return our connection object to the caller and clear it internally
+  // do not drop our reference - the caller now owns it.
+  MOZ_ASSERT(mConn);
+  return mConn.forget();
+}
+
+already_AddRefed<HttpConnectionBase> ConnectionHandle::HttpConnection() {
+  RefPtr<HttpConnectionBase> rv(mConn);
+  return rv.forget();
+}
+
+void ConnectionHandle::TopLevelOuterContentWindowIdChanged(uint64_t windowId) {
+  // Do nothing.
+}
+
+// nsConnectionEntry
+
+nsHttpConnectionMgr::nsConnectionEntry::nsConnectionEntry(
+    nsHttpConnectionInfo* ci)
+    : mConnInfo(ci),
+      mUsingSpdy(false),
+      mCanUseSpdy(true),
+      mPreferIPv4(false),
+      mPreferIPv6(false),
+      mUsedForConnection(false),
+      mDoNotDestroy(false) {
+  MOZ_COUNT_CTOR(nsConnectionEntry);
+
+  if (mConnInfo->FirstHopSSL() && !mConnInfo->IsHttp3()) {
+    mUseFastOpen = gHttpHandler->UseFastOpen();
+  } else {
+    // Only allow the TCP fast open on a secure connection.
+    mUseFastOpen = false;
+  }
+
+  LOG(("nsConnectionEntry::nsConnectionEntry this=%p key=%s", this,
+       ci->HashKey().get()));
+}
+
+bool nsHttpConnectionMgr::nsConnectionEntry::AvailableForDispatchNow() {
+  if (mIdleConns.Length() && mIdleConns[0]->CanReuse()) {
+    return true;
+  }
+
+  return gHttpHandler->ConnMgr()->GetH2orH3ActiveConn(this, false) ? true
+                                                                   : false;
+}
+
 bool nsHttpConnectionMgr::GetConnectionData(nsTArray<HttpRetParams>* aArg) {
   for (auto iter = mCT.Iter(); !iter.Done(); iter.Next()) {
-    RefPtr<ConnectionEntry> ent = iter.Data();
+    RefPtr<nsConnectionEntry> ent = iter.Data();
 
     if (ent->mConnInfo->GetPrivate()) {
       continue;
     }
-    aArg->AppendElement(ent->GetConnectionData());
+
+    HttpRetParams data;
+    data.host = ent->mConnInfo->Origin();
+    data.port = ent->mConnInfo->OriginPort();
+    for (uint32_t i = 0; i < ent->mActiveConns.Length(); i++) {
+      HttpConnInfo info;
+      RefPtr<nsHttpConnection> connTCP = do_QueryObject(ent->mActiveConns[i]);
+      if (connTCP) {
+        info.ttl = connTCP->TimeToLive();
+      } else {
+        info.ttl = 0;
+      }
+      info.rtt = ent->mActiveConns[i]->Rtt();
+      info.SetHTTPProtocolVersion(ent->mActiveConns[i]->Version());
+      data.active.AppendElement(info);
+    }
+    for (uint32_t i = 0; i < ent->mIdleConns.Length(); i++) {
+      HttpConnInfo info;
+      info.ttl = ent->mIdleConns[i]->TimeToLive();
+      info.rtt = ent->mIdleConns[i]->Rtt();
+      info.SetHTTPProtocolVersion(ent->mIdleConns[i]->Version());
+      data.idle.AppendElement(info);
+    }
+    for (uint32_t i = 0; i < ent->mHalfOpens.Length(); i++) {
+      HalfOpenSockets hSocket;
+      hSocket.speculative = ent->mHalfOpens[i]->IsSpeculative();
+      data.halfOpens.AppendElement(hSocket);
+    }
+    if (ent->mConnInfo->IsHttp3()) {
+      data.httpVersion = "HTTP/3"_ns;
+    } else if (ent->mUsingSpdy) {
+      data.httpVersion = "HTTP/2"_ns;
+    } else {
+      data.httpVersion = "HTTP <= 1.1"_ns;
+    }
+    data.ssl = ent->mConnInfo->EndToEndSSL();
+    aArg->AppendElement(data);
   }
 
   return true;
@@ -3348,16 +5529,62 @@ bool nsHttpConnectionMgr::GetConnectionData(nsTArray<HttpRetParams>* aArg) {
 
 void nsHttpConnectionMgr::ResetIPFamilyPreference(nsHttpConnectionInfo* ci) {
   MOZ_ASSERT(OnSocketThread(), "not on socket thread");
-  ConnectionEntry* ent = mCT.GetWeak(ci->HashKey());
+  nsConnectionEntry* ent = mCT.GetWeak(ci->HashKey());
   if (ent) {
     ent->ResetIPFamilyPreference();
+  }
+}
+
+uint32_t nsHttpConnectionMgr::nsConnectionEntry::UnconnectedHalfOpens() {
+  uint32_t unconnectedHalfOpens = 0;
+  for (uint32_t i = 0; i < mHalfOpens.Length(); ++i) {
+    if (!mHalfOpens[i]->HasConnected()) ++unconnectedHalfOpens;
+  }
+  return unconnectedHalfOpens;
+}
+
+void nsHttpConnectionMgr::nsConnectionEntry::RemoveHalfOpen(
+    nsHalfOpenSocket* halfOpen) {
+  // A failure to create the transport object at all
+  // will result in it not being present in the halfopen table. That's expected.
+  if (mHalfOpens.RemoveElement(halfOpen)) {
+    if (halfOpen->IsSpeculative()) {
+      Telemetry::AutoCounter<Telemetry::HTTPCONNMGR_UNUSED_SPECULATIVE_CONN>
+          unusedSpeculativeConn;
+      ++unusedSpeculativeConn;
+
+      if (halfOpen->IsFromPredictor()) {
+        Telemetry::AutoCounter<Telemetry::PREDICTOR_TOTAL_PRECONNECTS_UNUSED>
+            totalPreconnectsUnused;
+        ++totalPreconnectsUnused;
+      }
+    }
+
+    MOZ_ASSERT(gHttpHandler->ConnMgr()->mNumHalfOpenConns);
+    if (gHttpHandler->ConnMgr()->mNumHalfOpenConns) {  // just in case
+      gHttpHandler->ConnMgr()->mNumHalfOpenConns--;
+    }
+  } else {
+    mHalfOpenFastOpenBackups.RemoveElement(halfOpen);
+  }
+
+  if (!UnconnectedHalfOpens()) {
+    // perhaps this reverted RestrictConnections()
+    // use the PostEvent version of processpendingq to avoid
+    // altering the pending q vector from an arbitrary stack
+    nsresult rv = gHttpHandler->ConnMgr()->ProcessPendingQ(mConnInfo);
+    if (NS_FAILED(rv)) {
+      LOG(
+          ("nsHttpConnectionMgr::nsConnectionEntry::RemoveHalfOpen\n"
+           "    failed to process pending queue\n"));
+    }
   }
 }
 
 void nsHttpConnectionMgr::ExcludeHttp2(const nsHttpConnectionInfo* ci) {
   LOG(("nsHttpConnectionMgr::ExcludeHttp2 excluding ci %s",
        ci->HashKey().BeginReading()));
-  ConnectionEntry* ent = mCT.GetWeak(ci->HashKey());
+  nsConnectionEntry* ent = mCT.GetWeak(ci->HashKey());
   if (!ent) {
     LOG(("nsHttpConnectionMgr::ExcludeHttp2 no entry found?!"));
     return;
@@ -3369,13 +5596,177 @@ void nsHttpConnectionMgr::ExcludeHttp2(const nsHttpConnectionInfo* ci) {
 void nsHttpConnectionMgr::ExcludeHttp3(const nsHttpConnectionInfo* ci) {
   LOG(("nsHttpConnectionMgr::ExcludeHttp3 exclude ci %s",
        ci->HashKey().BeginReading()));
-  ConnectionEntry* ent = mCT.GetWeak(ci->HashKey());
+  nsConnectionEntry* ent = mCT.GetWeak(ci->HashKey());
   if (!ent) {
     LOG(("nsHttpConnectionMgr::ExcludeHttp3 no entry found?!"));
     return;
   }
 
   ent->DontReuseHttp3Conn();
+}
+
+void nsHttpConnectionMgr::nsConnectionEntry::DisallowHttp2() {
+  mCanUseSpdy = false;
+
+  // If we have any spdy connections, we want to go ahead and close them when
+  // they're done so we can free up some connections.
+  for (uint32_t i = 0; i < mActiveConns.Length(); ++i) {
+    if (mActiveConns[i]->UsingSpdy()) {
+      mActiveConns[i]->DontReuse();
+    }
+  }
+  for (uint32_t i = 0; i < mIdleConns.Length(); ++i) {
+    if (mIdleConns[i]->UsingSpdy()) {
+      mIdleConns[i]->DontReuse();
+    }
+  }
+
+  // Can't coalesce if we're not using spdy
+  mCoalescingKeys.Clear();
+}
+
+void nsHttpConnectionMgr::nsConnectionEntry::DontReuseHttp3Conn() {
+  MOZ_ASSERT(mConnInfo->IsHttp3());
+
+  // If we have any spdy connections, we want to go ahead and close them when
+  // they're done so we can free up some connections.
+  for (uint32_t i = 0; i < mActiveConns.Length(); ++i) {
+    mActiveConns[i]->DontReuse();
+  }
+
+  // Can't coalesce if we're not using http3
+  mCoalescingKeys.Clear();
+}
+
+void nsHttpConnectionMgr::nsConnectionEntry::RecordIPFamilyPreference(
+    uint16_t family) {
+  LOG(("nsConnectionEntry::RecordIPFamilyPreference %p, af=%u", this, family));
+
+  if (family == PR_AF_INET && !mPreferIPv6) {
+    mPreferIPv4 = true;
+  }
+
+  if (family == PR_AF_INET6 && !mPreferIPv4) {
+    mPreferIPv6 = true;
+  }
+
+  LOG(("  %p prefer ipv4=%d, ipv6=%d", this, (bool)mPreferIPv4,
+       (bool)mPreferIPv6));
+}
+
+void nsHttpConnectionMgr::nsConnectionEntry::ResetIPFamilyPreference() {
+  LOG(("nsConnectionEntry::ResetIPFamilyPreference %p", this));
+
+  mPreferIPv4 = false;
+  mPreferIPv6 = false;
+}
+
+bool net::nsHttpConnectionMgr::nsConnectionEntry::PreferenceKnown() const {
+  return (bool)mPreferIPv4 || (bool)mPreferIPv6;
+}
+
+size_t nsHttpConnectionMgr::nsConnectionEntry::PendingQLength() const {
+  size_t length = 0;
+  for (auto it = mPendingTransactionTable.ConstIter(); !it.Done(); it.Next()) {
+    length += it.UserData()->Length();
+  }
+
+  return length;
+}
+
+void nsHttpConnectionMgr::nsConnectionEntry::InsertTransaction(
+    PendingTransactionInfo* info,
+    bool aInsertAsFirstForTheSamePriority /*= false*/) {
+  LOG(
+      ("nsHttpConnectionMgr::nsConnectionEntry::InsertTransaction"
+       " trans=%p, windowId=%" PRIu64 "\n",
+       info->mTransaction.get(),
+       info->mTransaction->TopLevelOuterContentWindowId()));
+
+  uint64_t windowId = TabIdForQueuing(info->mTransaction);
+  nsTArray<RefPtr<PendingTransactionInfo>>* infoArray;
+  if (!mPendingTransactionTable.Get(windowId, &infoArray)) {
+    infoArray = new nsTArray<RefPtr<PendingTransactionInfo>>();
+    mPendingTransactionTable.Put(windowId, infoArray);
+  }
+
+  gHttpHandler->ConnMgr()->InsertTransactionSorted(
+      *infoArray, info, aInsertAsFirstForTheSamePriority);
+}
+
+void nsHttpConnectionMgr::nsConnectionEntry::AppendPendingQForFocusedWindow(
+    uint64_t windowId, nsTArray<RefPtr<PendingTransactionInfo>>& result,
+    uint32_t maxCount) {
+  nsTArray<RefPtr<PendingTransactionInfo>>* infoArray = nullptr;
+  if (!mPendingTransactionTable.Get(windowId, &infoArray)) {
+    result.Clear();
+    return;
+  }
+
+  uint32_t countToAppend = maxCount;
+  countToAppend = countToAppend > infoArray->Length() || countToAppend == 0
+                      ? infoArray->Length()
+                      : countToAppend;
+
+  result.InsertElementsAt(result.Length(), infoArray->Elements(),
+                          countToAppend);
+  infoArray->RemoveElementsAt(0, countToAppend);
+
+  LOG(
+      ("nsConnectionEntry::AppendPendingQForFocusedWindow [ci=%s], "
+       "pendingQ count=%zu window.count=%zu for focused window (id=%" PRIu64
+       ")\n",
+       mConnInfo->HashKey().get(), result.Length(), infoArray->Length(),
+       windowId));
+}
+
+void nsHttpConnectionMgr::nsConnectionEntry::AppendPendingQForNonFocusedWindows(
+    uint64_t windowId, nsTArray<RefPtr<PendingTransactionInfo>>& result,
+    uint32_t maxCount) {
+  // XXX Adjust the order of transactions in a smarter manner.
+  uint32_t totalCount = 0;
+  for (auto it = mPendingTransactionTable.Iter(); !it.Done(); it.Next()) {
+    if (windowId && it.Key() == windowId) {
+      continue;
+    }
+
+    uint32_t count = 0;
+    for (; count < it.UserData()->Length(); ++count) {
+      if (maxCount && totalCount == maxCount) {
+        break;
+      }
+
+      // Because elements in |result| could come from multiple penndingQ,
+      // call |InsertTransactionSorted| to make sure the order is correct.
+      gHttpHandler->ConnMgr()->InsertTransactionSorted(
+          result, it.UserData()->ElementAt(count));
+      ++totalCount;
+    }
+    it.UserData()->RemoveElementsAt(0, count);
+
+    if (maxCount && totalCount == maxCount) {
+      if (it.UserData()->Length()) {
+        // There are still some pending transactions for background
+        // tabs but we limit their dispatch.  This is considered as
+        // an active tab optimization.
+        nsHttp::NotifyActiveTabLoadOptimization();
+      }
+      break;
+    }
+  }
+
+  LOG(
+      ("nsConnectionEntry::AppendPendingQForNonFocusedWindows [ci=%s], "
+       "pendingQ count=%zu for non focused window\n",
+       mConnInfo->HashKey().get(), result.Length()));
+}
+
+void nsHttpConnectionMgr::nsConnectionEntry::RemoveEmptyPendingQ() {
+  for (auto it = mPendingTransactionTable.Iter(); !it.Done(); it.Next()) {
+    if (it.UserData()->IsEmpty()) {
+      it.Remove();
+    }
+  }
 }
 
 void nsHttpConnectionMgr::MoveToWildCardConnEntry(
@@ -3389,7 +5780,7 @@ void nsHttpConnectionMgr::MoveToWildCardConnEntry(
        "change CI from %s to %s\n",
        proxyConn, specificCI->HashKey().get(), wildCardCI->HashKey().get()));
 
-  ConnectionEntry* ent = mCT.GetWeak(specificCI->HashKey());
+  nsConnectionEntry* ent = mCT.GetWeak(specificCI->HashKey());
   LOG(
       ("nsHttpConnectionMgr::MakeConnEntryWildCard conn %p using ent %p (spdy "
        "%d)\n",
@@ -3399,7 +5790,8 @@ void nsHttpConnectionMgr::MoveToWildCardConnEntry(
     return;
   }
 
-  ConnectionEntry* wcEnt = GetOrCreateConnectionEntry(wildCardCI, true, false);
+  nsConnectionEntry* wcEnt =
+      GetOrCreateConnectionEntry(wildCardCI, true, false);
   if (wcEnt == ent) {
     // nothing to do!
     return;
@@ -3409,16 +5801,36 @@ void nsHttpConnectionMgr::MoveToWildCardConnEntry(
   LOG(
       ("nsHttpConnectionMgr::MakeConnEntryWildCard ent %p "
        "idle=%zu active=%zu half=%zu pending=%zu\n",
-       ent, ent->IdleConnectionsLength(), ent->ActiveConnsLength(),
-       ent->HalfOpensLength(), ent->PendingQueueLength()));
+       ent, ent->mIdleConns.Length(), ent->mActiveConns.Length(),
+       ent->mHalfOpens.Length(), ent->PendingQLength()));
 
   LOG(
       ("nsHttpConnectionMgr::MakeConnEntryWildCard wc-ent %p "
        "idle=%zu active=%zu half=%zu pending=%zu\n",
-       wcEnt, wcEnt->IdleConnectionsLength(), wcEnt->ActiveConnsLength(),
-       wcEnt->HalfOpensLength(), wcEnt->PendingQueueLength()));
+       wcEnt, wcEnt->mIdleConns.Length(), wcEnt->mActiveConns.Length(),
+       wcEnt->mHalfOpens.Length(), wcEnt->PendingQLength()));
 
-  ent->MoveConnection(proxyConn, wcEnt);
+  int32_t count = ent->mActiveConns.Length();
+  RefPtr<HttpConnectionBase> deleteProtector(proxyConn);
+  for (int32_t i = 0; i < count; ++i) {
+    if (ent->mActiveConns[i] == proxyConn) {
+      ent->mActiveConns.RemoveElementAt(i);
+      wcEnt->mActiveConns.InsertElementAt(0, proxyConn);
+      return;
+    }
+  }
+
+  RefPtr<nsHttpConnection> proxyConnTCP = do_QueryObject(proxyConn);
+  if (proxyConnTCP) {
+    count = ent->mIdleConns.Length();
+    for (int32_t i = 0; i < count; ++i) {
+      if (ent->mIdleConns[i] == proxyConnTCP) {
+        ent->mIdleConns.RemoveElementAt(i);
+        wcEnt->mIdleConns.InsertElementAt(0, proxyConnTCP);
+        return;
+      }
+    }
+  }
 }
 
 bool nsHttpConnectionMgr::MoveTransToHTTPSSVCConnEntry(
@@ -3431,10 +5843,10 @@ bool nsHttpConnectionMgr::MoveTransToHTTPSSVCConnEntry(
   bool prohibitWildCard = !!aTrans->TunnelProvider();
   bool noHttp3 = aTrans->Caps() & NS_HTTP_DISALLOW_HTTP3;
   // Step 1: Check if the new entry is the same as the old one.
-  ConnectionEntry* oldEntry = GetOrCreateConnectionEntry(
+  nsConnectionEntry* oldEntry = GetOrCreateConnectionEntry(
       aTrans->ConnectionInfo(), prohibitWildCard, noHttp3);
 
-  ConnectionEntry* newEntry =
+  nsConnectionEntry* newEntry =
       GetOrCreateConnectionEntry(aNewCI, prohibitWildCard, noHttp3);
 
   if (oldEntry == newEntry) {
@@ -3446,7 +5858,7 @@ bool nsHttpConnectionMgr::MoveTransToHTTPSSVCConnEntry(
   // We will abandon all half-open sockets belonging to the given
   // transaction.
   nsTArray<RefPtr<PendingTransactionInfo>>* infoArray =
-      oldEntry->GetTransactionPendingQHelper(aTrans);
+      GetTransactionPendingQHelper(oldEntry, aTrans);
 
   RefPtr<PendingTransactionInfo> pendingTransInfo;
   transIndex =
@@ -3461,10 +5873,15 @@ bool nsHttpConnectionMgr::MoveTransToHTTPSSVCConnEntry(
     return false;
   }
 
-  MOZ_ASSERT(pendingTransInfo->Transaction() == aTrans);
+  MOZ_ASSERT(pendingTransInfo->mTransaction == aTrans);
 
   // Abandon all half-open sockets belonging to the given transaction.
-  pendingTransInfo->AbandonHalfOpenAndForgetActiveConn();
+  RefPtr<nsHalfOpenSocket> half = do_QueryReferent(pendingTransInfo->mHalfOpen);
+  if (half) {
+    half->Abandon();
+  }
+  pendingTransInfo->mHalfOpen = nullptr;
+  pendingTransInfo->mActiveConn = nullptr;
 
   // Step 3: Add the transaction to the new entry.
   aTrans->UpdateConnectionInfo(aNewCI);
@@ -3472,55 +5889,10 @@ bool nsHttpConnectionMgr::MoveTransToHTTPSSVCConnEntry(
   return true;
 }
 
-void nsHttpConnectionMgr::IncreaseNumHalfOpenConns() { mNumHalfOpenConns++; }
-
-void nsHttpConnectionMgr::DecreaseNumHalfOpenConns() {
-  MOZ_ASSERT(mNumHalfOpenConns);
-  if (mNumHalfOpenConns) {  // just in case
-    mNumHalfOpenConns--;
-  }
-}
-
-already_AddRefed<PendingTransactionInfo>
-nsHttpConnectionMgr::FindTransactionHelper(bool removeWhenFound,
-                                           ConnectionEntry* aEnt,
-                                           nsAHttpTransaction* aTrans) {
-  nsTArray<RefPtr<PendingTransactionInfo>>* pendingQ =
-      aEnt->GetTransactionPendingQHelper(aTrans);
-
-  int32_t index =
-      pendingQ ? pendingQ->IndexOf(aTrans, 0, PendingComparator()) : -1;
-
-  RefPtr<PendingTransactionInfo> info;
-  if (index != -1) {
-    info = (*pendingQ)[index];
-    if (removeWhenFound) {
-      pendingQ->RemoveElementAt(index);
-    }
-  }
-  return info.forget();
-}
-
 nsHttpConnectionMgr* nsHttpConnectionMgr::AsHttpConnectionMgr() { return this; }
 
 HttpConnectionMgrParent* nsHttpConnectionMgr::AsHttpConnectionMgrParent() {
   return nullptr;
-}
-
-void nsHttpConnectionMgr::NewIdleConnectionAdded(uint32_t timeToLive) {
-  mNumIdleConns++;
-
-  // If the added connection was first idle connection or has shortest
-  // time to live among the watched connections, pruning dead
-  // connections needs to be done when it can't be reused anymore.
-  if (!mTimer || NowInSeconds() + timeToLive < mTimeOfNextWakeUp)
-    PruneDeadConnectionsAfter(timeToLive);
-}
-
-void nsHttpConnectionMgr::DecrementNumIdleConns() {
-  MOZ_ASSERT(mNumIdleConns);
-  mNumIdleConns--;
-  ConditionallyStopPruneDeadConnectionsTimer();
 }
 
 }  // namespace net
